feat(plugin)!: delete PluginHostService god-service; rehoming complete (holomush-eykuh.1)
Relocate the 22 authoritative host-RPC handler bodies + settings helpers from
pluginHostServiceServer onto the per-capability host.v1 servers (inverting the
T3 delegation), delete pluginHostServiceServer + the PluginHostService proto
block + orphaned messages, regenerate, and add the rehoming-completeness
meta-test. No behavior change.
