+++ pkg/proto/holomush/plugin/v1/pluginv1connect/plugin.connect.go
@@ -28,8 +28,6 @@ const _ = connect.IsAtLeastVersion1_13_0
 const (
 	// PluginServiceName is the fully-qualified name of the PluginService service.
 	PluginServiceName = "holomush.plugin.v1.PluginService"
-	// PluginHostServiceName is the fully-qualified name of the PluginHostService service.
-	PluginHostServiceName = "holomush.plugin.v1.PluginHostService"
 )
 
 // These constants are the fully-qualified names of the RPCs defined in this package. They're
@@ -51,72 +49,6 @@ const (
 	// PluginServiceQuerySessionStreamsProcedure is the fully-qualified name of the PluginService's
 	// QuerySessionStreams RPC.
 	PluginServiceQuerySessionStreamsProcedure = "/holomush.plugin.v1.PluginService/QuerySessionStreams"
-	// PluginHostServiceEmitEventProcedure is the fully-qualified name of the PluginHostService's
-	// EmitEvent RPC.
-	PluginHostServiceEmitEventProcedure = "/holomush.plugin.v1.PluginHostService/EmitEvent"
-	// PluginHostServiceLogProcedure is the fully-qualified name of the PluginHostService's Log RPC.
-	PluginHostServiceLogProcedure = "/holomush.plugin.v1.PluginHostService/Log"
-	// PluginHostServiceKVGetProcedure is the fully-qualified name of the PluginHostService's KVGet RPC.
-	PluginHostServiceKVGetProcedure = "/holomush.plugin.v1.PluginHostService/KVGet"
-	// PluginHostServiceKVSetProcedure is the fully-qualified name of the PluginHostService's KVSet RPC.
-	PluginHostServiceKVSetProcedure = "/holomush.plugin.v1.PluginHostService/KVSet"
-	// PluginHostServiceKVDeleteProcedure is the fully-qualified name of the PluginHostService's
-	// KVDelete RPC.
-	PluginHostServiceKVDeleteProcedure = "/holomush.plugin.v1.PluginHostService/KVDelete"
-	// PluginHostServiceAddSessionStreamProcedure is the fully-qualified name of the PluginHostService's
-	// AddSessionStream RPC.
-	PluginHostServiceAddSessionStreamProcedure = "/holomush.plugin.v1.PluginHostService/AddSessionStream"
-	// PluginHostServiceRemoveSessionStreamProcedure is the fully-qualified name of the
-	// PluginHostService's RemoveSessionStream RPC.
-	PluginHostServiceRemoveSessionStreamProcedure = "/holomush.plugin.v1.PluginHostService/RemoveSessionStream"
-	// PluginHostServiceJoinFocusProcedure is the fully-qualified name of the PluginHostService's
-	// JoinFocus RPC.
-	PluginHostServiceJoinFocusProcedure = "/holomush.plugin.v1.PluginHostService/JoinFocus"
-	// PluginHostServiceLeaveFocusProcedure is the fully-qualified name of the PluginHostService's
-	// LeaveFocus RPC.
-	PluginHostServiceLeaveFocusProcedure = "/holomush.plugin.v1.PluginHostService/LeaveFocus"
-	// PluginHostServiceLeaveFocusByTargetProcedure is the fully-qualified name of the
-	// PluginHostService's LeaveFocusByTarget RPC.
-	PluginHostServiceLeaveFocusByTargetProcedure = "/holomush.plugin.v1.PluginHostService/LeaveFocusByTarget"
-	// PluginHostServicePresentFocusProcedure is the fully-qualified name of the PluginHostService's
-	// PresentFocus RPC.
-	PluginHostServicePresentFocusProcedure = "/holomush.plugin.v1.PluginHostService/PresentFocus"
-	// PluginHostServiceQueryStreamHistoryProcedure is the fully-qualified name of the
-	// PluginHostService's QueryStreamHistory RPC.
-	PluginHostServiceQueryStreamHistoryProcedure = "/holomush.plugin.v1.PluginHostService/QueryStreamHistory"
-	// PluginHostServiceDecryptOwnAuditRowsProcedure is the fully-qualified name of the
-	// PluginHostService's DecryptOwnAuditRows RPC.
-	PluginHostServiceDecryptOwnAuditRowsProcedure = "/holomush.plugin.v1.PluginHostService/DecryptOwnAuditRows"
-	// PluginHostServiceRequestEmitTokenProcedure is the fully-qualified name of the PluginHostService's
-	// RequestEmitToken RPC.
-	PluginHostServiceRequestEmitTokenProcedure = "/holomush.plugin.v1.PluginHostService/RequestEmitToken"
-	// PluginHostServiceSetConnectionFocusProcedure is the fully-qualified name of the
-	// PluginHostService's SetConnectionFocus RPC.
-	PluginHostServiceSetConnectionFocusProcedure = "/holomush.plugin.v1.PluginHostService/SetConnectionFocus"
-	// PluginHostServiceGetConnectionFocusProcedure is the fully-qualified name of the
-	// PluginHostService's GetConnectionFocus RPC.
-	PluginHostServiceGetConnectionFocusProcedure = "/holomush.plugin.v1.PluginHostService/GetConnectionFocus"
-	// PluginHostServiceAutoFocusOnJoinProcedure is the fully-qualified name of the PluginHostService's
-	// AutoFocusOnJoin RPC.
-	PluginHostServiceAutoFocusOnJoinProcedure = "/holomush.plugin.v1.PluginHostService/AutoFocusOnJoin"
-	// PluginHostServiceIsAnyConnFocusedProcedure is the fully-qualified name of the PluginHostService's
-	// IsAnyConnFocused RPC.
-	PluginHostServiceIsAnyConnFocusedProcedure = "/holomush.plugin.v1.PluginHostService/IsAnyConnFocused"
-	// PluginHostServiceEvaluateProcedure is the fully-qualified name of the PluginHostService's
-	// Evaluate RPC.
-	PluginHostServiceEvaluateProcedure = "/holomush.plugin.v1.PluginHostService/Evaluate"
-	// PluginHostServiceListCommandsProcedure is the fully-qualified name of the PluginHostService's
-	// ListCommands RPC.
-	PluginHostServiceListCommandsProcedure = "/holomush.plugin.v1.PluginHostService/ListCommands"
-	// PluginHostServiceGetCommandHelpProcedure is the fully-qualified name of the PluginHostService's
-	// GetCommandHelp RPC.
-	PluginHostServiceGetCommandHelpProcedure = "/holomush.plugin.v1.PluginHostService/GetCommandHelp"
-	// PluginHostServiceGetSettingProcedure is the fully-qualified name of the PluginHostService's
-	// GetSetting RPC.
-	PluginHostServiceGetSettingProcedure = "/holomush.plugin.v1.PluginHostService/GetSetting"
-	// PluginHostServiceSetSettingProcedure is the fully-qualified name of the PluginHostService's
-	// SetSetting RPC.
-	PluginHostServiceSetSettingProcedure = "/holomush.plugin.v1.PluginHostService/SetSetting"
 )
 
 // PluginServiceClient is a client for the holomush.plugin.v1.PluginService service.
@@ -309,887 +241,4 @@ func (UnimplementedPluginServiceHandler) HandleCommand(context.Context, *connect
 
 func (UnimplementedPluginServiceHandler) QuerySessionStreams(context.Context, *connect.Request[v1.QuerySessionStreamsRequest]) (*connect.Response[v1.QuerySessionStreamsResponse], error) {
 	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginService.QuerySessionStreams is not implemented"))
-}
-
-// PluginHostServiceClient is a client for the holomush.plugin.v1.PluginHostService service.
-type PluginHostServiceClient interface {
-	// EmitEvent publishes one plugin-originated event onto the bus through the
-	// host emit fence. SERVED: pluginHostServiceServer.EmitEvent. The caller's
-	// identity is NOT trusted from the wire — the plugin presents a host-issued
-	// dispatch token in the x-holomush-emit-token metadata header, the host
-	// recovers the vouched-for actor from tokenStore.Lookup(pluginName, token),
-	// and a missing/foreign token is rejected (EMIT_TOKEN_MISSING /
-	// EMIT_TOKEN_REJECTED). The recovered actor then flows through
-	// PluginEventEmitter.Emit, which enforces the manifest gates: emits (subject
-	// namespace must be declared), actor_kinds_claimable (actor kind must be
-	// listed — EMIT_ACTOR_KIND_NOT_CLAIMABLE), and the crypto.emits sensitivity
-	// fence. These gates fire identically for Lua and binary plugins (plugin
-	// runtime symmetry); the token mechanism is the binary-side forgery fence.
-	EmitEvent(context.Context, *connect.Request[v1.PluginHostServiceEmitEventRequest]) (*connect.Response[v1.PluginHostServiceEmitEventResponse], error)
-	// Log forwards a plugin log line to the host logger. DECLARED BUT UNSERVED:
-	// pluginHostServiceServer does not implement it and no production client
-	// calls it, so it currently returns codes.Unimplemented (holomush-l6std).
-	Log(context.Context, *connect.Request[v1.PluginHostServiceLogRequest]) (*connect.Response[v1.PluginHostServiceLogResponse], error)
-	// KVGet reads a value from the plugin's namespaced key-value store.
-	// DECLARED BUT UNSERVED (holomush-l6std): no server impl, no client; returns
-	// codes.Unimplemented today.
-	KVGet(context.Context, *connect.Request[v1.PluginHostServiceKVGetRequest]) (*connect.Response[v1.PluginHostServiceKVGetResponse], error)
-	// KVSet writes a value into the plugin's namespaced key-value store.
-	// DECLARED BUT UNSERVED (holomush-l6std): no server impl, no client; returns
-	// codes.Unimplemented today.
-	KVSet(context.Context, *connect.Request[v1.PluginHostServiceKVSetRequest]) (*connect.Response[v1.PluginHostServiceKVSetResponse], error)
-	// KVDelete removes a key from the plugin's namespaced key-value store.
-	// DECLARED BUT UNSERVED (holomush-l6std): no server impl, no client; returns
-	// codes.Unimplemented today.
-	KVDelete(context.Context, *connect.Request[v1.PluginHostServiceKVDeleteRequest]) (*connect.Response[v1.PluginHostServiceKVDeleteResponse], error)
-	// AddSessionStream subscribes an active session to an additional stream
-	// mid-session. DECLARED BUT UNSERVED (holomush-l6std): no server impl, no
-	// production client; returns codes.Unimplemented today. (The wire comment
-	// about SESSION_NOT_FOUND describes the intended-but-unimplemented contract.)
-	AddSessionStream(context.Context, *connect.Request[v1.PluginHostServiceAddSessionStreamRequest]) (*connect.Response[v1.PluginHostServiceAddSessionStreamResponse], error)
-	// RemoveSessionStream unsubscribes an active session from a stream.
-	// DECLARED BUT UNSERVED (holomush-l6std): no server impl, no production
-	// client; returns codes.Unimplemented today.
-	RemoveSessionStream(context.Context, *connect.Request[v1.PluginHostServiceRemoveSessionStreamRequest]) (*connect.Response[v1.PluginHostServiceRemoveSessionStreamResponse], error)
-	// JoinFocus adds a focus membership (e.g. a scene) to a session via the
-	// host focus coordinator. SERVED: pluginHostServiceServer.JoinFocus. The
-	// plugin declares intent; the coordinator applies the kind-specific replay
-	// policy. Fails if the focus coordinator is not configured.
-	JoinFocus(context.Context, *connect.Request[v1.PluginHostServiceJoinFocusRequest]) (*connect.Response[v1.PluginHostServiceJoinFocusResponse], error)
-	// LeaveFocus removes one focus membership from a session. SERVED:
-	// pluginHostServiceServer.LeaveFocus. Idempotent — leaving a target the
-	// session does not hold is a successful no-op.
-	LeaveFocus(context.Context, *connect.Request[v1.PluginHostServiceLeaveFocusRequest]) (*connect.Response[v1.PluginHostServiceLeaveFocusResponse], error)
-	// LeaveFocusByTarget removes the given focus membership from every
-	// non-expired session that holds it — cross-session fan-out (e.g. a
-	// scene-end reaching all participants). SERVED:
-	// pluginHostServiceServer.LeaveFocusByTarget. Partial success is normal:
-	// per-session failures are aggregated into the response (succeeded /
-	// total_scanned / failed_session_ids) rather than aborting the sweep; the
-	// RPC error is reserved for the enumeration step itself failing.
-	LeaveFocusByTarget(context.Context, *connect.Request[v1.PluginHostServiceLeaveFocusByTargetRequest]) (*connect.Response[v1.PluginHostServiceLeaveFocusByTargetResponse], error)
-	// PresentFocus repoints a session's PresentingFocus to an existing
-	// membership. SERVED: pluginHostServiceServer.PresentFocus. The target MUST
-	// already be in the session's FocusMemberships (membership is validated, not
-	// implicitly created).
-	PresentFocus(context.Context, *connect.Request[v1.PluginHostServicePresentFocusRequest]) (*connect.Response[v1.PluginHostServicePresentFocusResponse], error)
-	// QueryStreamHistory reads the tail of a stream for plugin-side display,
-	// backward-paginated by opaque cursor. SERVED:
-	// pluginHostServiceServer.QueryStreamHistory via HistoryReader.ReplayTail.
-	// Read-only: it does not advance session cursors or mutate session state.
-	// A negative count is rejected (INVALID_ARGUMENT); count is CLAMPED to
-	// maxQueryStreamHistoryCount (500), not rejected, when too large.
-	QueryStreamHistory(context.Context, *connect.Request[v1.PluginHostServiceQueryStreamHistoryRequest]) (*connect.Response[v1.PluginHostServiceQueryStreamHistoryResponse], error)
-	// DecryptOwnAuditRows decrypts a batch of the calling plugin's OWN encrypted
-	// audit rows host-side; the plugin never holds a DEK. SERVED:
-	// pluginHostServiceServer.DecryptOwnAuditRows via ReadbackDecryptor.
-	// DecryptOwnRows. Authorization is two-gate (INV-CRYPTO-27): OwnerMap subject
-	// ownership (g1) plus the crypto.emits[].readback manifest flag (g2). Each
-	// input row gets an independent RowResult (INV-CRYPTO-37) carrying either
-	// plaintext or a stable snake_case no_plaintext_reason ("not_owner",
-	// "auth_guard_deny", "dek_missing", "downgrade_refused", "stale_dek",
-	// "audit_queue_full", "internal"; readback.go reasonToWire). Request /
-	// response shapes and RowResult live in audit.proto (AuditRow domain).
-	DecryptOwnAuditRows(context.Context, *connect.Request[v1.DecryptOwnAuditRowsRequest]) (*connect.Response[v1.DecryptOwnAuditRowsResponse], error)
-	// RequestEmitToken issues a self-token bound to {ActorPlugin, pluginName} so
-	// a plugin-served gRPC handler (e.g. SceneService.CreateScene) — which is NOT
-	// reached via DeliverEvent/DeliverCommand and so holds no dispatch token —
-	// can still call EmitEvent. SERVED: pluginHostServiceServer.RequestEmitToken.
-	// The plugin's identity is taken from the mTLS-bound server struct
-	// (s.pluginName); the request carries no identity fields, so a plugin cannot
-	// impersonate another actor or escalate to a character actor through this
-	// RPC. The actor_kinds_claimable manifest gate still fires when the issued
-	// token is later spent at EmitEvent. (Spec §3.3.5 / §5.4 two-token pattern.)
-	RequestEmitToken(context.Context, *connect.Request[v1.PluginHostServiceRequestEmitTokenRequest]) (*connect.Response[v1.PluginHostServiceRequestEmitTokenResponse], error)
-	// SetConnectionFocus is the Phase-5 explicit focus mutation for a single
-	// Connection. SERVED: pluginHostServiceServer.SetConnectionFocus. The
-	// substrate validates the requested membership against the session's
-	// FocusMemberships (D4), then writes Connection.FocusKey and (D9-gated)
-	// Info.PresentingFocus atomically under one Store-lock acquisition (D7).
-	SetConnectionFocus(context.Context, *connect.Request[v1.PluginHostServiceSetConnectionFocusRequest]) (*connect.Response[v1.PluginHostServiceSetConnectionFocusResponse], error)
-	// GetConnectionFocus returns the named connection's current per-connection
-	// focus, or absent when the connection is grid-focused (FocusKey nil) or
-	// unknown. Read-only counterpart of SetConnectionFocus; lets plugins route
-	// connection-scoped operations (e.g. scene pose) to the focused target.
-	// See goplugin/host_service.go::GetConnectionFocus.
-	GetConnectionFocus(context.Context, *connect.Request[v1.PluginHostServiceGetConnectionFocusRequest]) (*connect.Response[v1.PluginHostServiceGetConnectionFocusResponse], error)
-	// AutoFocusOnJoin is the Phase-5 fan-out that focuses all of a character's
-	// terminal/telnet connections on a scene at once. SERVED:
-	// pluginHostServiceServer.AutoFocusOnJoin. Connections already explicitly
-	// focused elsewhere are skipped (D8). The caller MUST have completed
-	// JoinFocus first, since the substrate requires the membership to exist.
-	AutoFocusOnJoin(context.Context, *connect.Request[v1.PluginHostServiceAutoFocusOnJoinRequest]) (*connect.Response[v1.PluginHostServiceAutoFocusOnJoinResponse], error)
-	// IsAnyConnFocused is the Phase-5 notification-emission helper: it reports
-	// whether any of the character's connections currently focuses the given
-	// scene, so callers can decide whether to emit a focus-related notification.
-	// SERVED: pluginHostServiceServer.IsAnyConnFocused.
-	IsAnyConnFocused(context.Context, *connect.Request[v1.PluginHostServiceIsAnyConnFocusedRequest]) (*connect.Response[v1.PluginHostServiceIsAnyConnFocusedResponse], error)
-	// Evaluate runs the host ABAC engine for one action against one resource
-	// instance owned by the calling plugin. SERVED:
-	// pluginHostServiceServer.Evaluate. The subject is derived host-side from the
-	// dispatch token exactly as EmitEvent does (token→actor recovery) — there is
-	// no subject field on the wire (spec §2, INV-PLUGIN-22). Fails closed on nil engine,
-	// missing/rejected token, empty actor subject, or a resource type the plugin
-	// does not own.
-	Evaluate(context.Context, *connect.Request[v1.PluginHostServiceEvaluateRequest]) (*connect.Response[v1.PluginHostServiceEvaluateResponse], error)
-	// ListCommands enumerates the commands the named character may execute,
-	// ABAC-filtered by the host. SERVED: pluginHostServiceServer.ListCommands,
-	// delegating to commandquery.Querier.Available. The subject is the request's
-	// character_id (parity with the Lua holomush.list_commands(character_id) host
-	// function — not the dispatch-token actor, since this is read-only metadata,
-	// not an actor-gated mutation). incomplete is true when engine errors hid
-	// some commands.
-	ListCommands(context.Context, *connect.Request[v1.PluginHostServiceListCommandsRequest]) (*connect.Response[v1.PluginHostServiceListCommandsResponse], error)
-	// GetCommandHelp returns full help detail for one command after an access
-	// check for character_id. SERVED: pluginHostServiceServer.GetCommandHelp,
-	// delegating to commandquery.Querier.Help. Mirrors the Lua
-	// holomush.get_command_help(name, character_id) host function.
-	GetCommandHelp(context.Context, *connect.Request[v1.PluginHostServiceGetCommandHelpRequest]) (*connect.Response[v1.PluginHostServiceGetCommandHelpResponse], error)
-	// GetSetting reads a single-scope setting in the calling plugin's owner
-	// partition (owner bound host-side from the authenticated plugin name, never
-	// from the request). The handler resolves SettingScope to its backing store;
-	// a missing key returns a successful response with found=false, never a
-	// codes.NotFound status error.
-	GetSetting(context.Context, *connect.Request[v1.PluginHostServiceGetSettingRequest]) (*connect.Response[v1.PluginHostServiceGetSettingResponse], error)
-	// SetSetting writes a single-scope setting in the calling plugin's partition;
-	// GAME scope requires an operator authorization decision (host-enforced, not
-	// trusted from the wire). The owner partition is bound host-side from the
-	// authenticated plugin name.
-	SetSetting(context.Context, *connect.Request[v1.PluginHostServiceSetSettingRequest]) (*connect.Response[v1.PluginHostServiceSetSettingResponse], error)
-}
-
-// NewPluginHostServiceClient constructs a client for the holomush.plugin.v1.PluginHostService
-// service. By default, it uses the Connect protocol with the binary Protobuf Codec, asks for
-// gzipped responses, and sends uncompressed requests. To use the gRPC or gRPC-Web protocols, supply
-// the connect.WithGRPC() or connect.WithGRPCWeb() options.
-//
-// The URL supplied here should be the base URL for the Connect or gRPC server (for example,
-// http://api.acme.com or https://acme.com/grpc).
-func NewPluginHostServiceClient(httpClient connect.HTTPClient, baseURL string, opts ...connect.ClientOption) PluginHostServiceClient {
-	baseURL = strings.TrimRight(baseURL, "/")
-	pluginHostServiceMethods := v1.File_holomush_plugin_v1_plugin_proto.Services().ByName("PluginHostService").Methods()
-	return &pluginHostServiceClient{
-		emitEvent: connect.NewClient[v1.PluginHostServiceEmitEventRequest, v1.PluginHostServiceEmitEventResponse](
-			httpClient,
-			baseURL+PluginHostServiceEmitEventProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("EmitEvent")),
-			connect.WithClientOptions(opts...),
-		),
-		log: connect.NewClient[v1.PluginHostServiceLogRequest, v1.PluginHostServiceLogResponse](
-			httpClient,
-			baseURL+PluginHostServiceLogProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("Log")),
-			connect.WithClientOptions(opts...),
-		),
-		kVGet: connect.NewClient[v1.PluginHostServiceKVGetRequest, v1.PluginHostServiceKVGetResponse](
-			httpClient,
-			baseURL+PluginHostServiceKVGetProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("KVGet")),
-			connect.WithClientOptions(opts...),
-		),
-		kVSet: connect.NewClient[v1.PluginHostServiceKVSetRequest, v1.PluginHostServiceKVSetResponse](
-			httpClient,
-			baseURL+PluginHostServiceKVSetProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("KVSet")),
-			connect.WithClientOptions(opts...),
-		),
-		kVDelete: connect.NewClient[v1.PluginHostServiceKVDeleteRequest, v1.PluginHostServiceKVDeleteResponse](
-			httpClient,
-			baseURL+PluginHostServiceKVDeleteProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("KVDelete")),
-			connect.WithClientOptions(opts...),
-		),
-		addSessionStream: connect.NewClient[v1.PluginHostServiceAddSessionStreamRequest, v1.PluginHostServiceAddSessionStreamResponse](
-			httpClient,
-			baseURL+PluginHostServiceAddSessionStreamProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("AddSessionStream")),
-			connect.WithClientOptions(opts...),
-		),
-		removeSessionStream: connect.NewClient[v1.PluginHostServiceRemoveSessionStreamRequest, v1.PluginHostServiceRemoveSessionStreamResponse](
-			httpClient,
-			baseURL+PluginHostServiceRemoveSessionStreamProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("RemoveSessionStream")),
-			connect.WithClientOptions(opts...),
-		),
-		joinFocus: connect.NewClient[v1.PluginHostServiceJoinFocusRequest, v1.PluginHostServiceJoinFocusResponse](
-			httpClient,
-			baseURL+PluginHostServiceJoinFocusProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("JoinFocus")),
-			connect.WithClientOptions(opts...),
-		),
-		leaveFocus: connect.NewClient[v1.PluginHostServiceLeaveFocusRequest, v1.PluginHostServiceLeaveFocusResponse](
-			httpClient,
-			baseURL+PluginHostServiceLeaveFocusProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("LeaveFocus")),
-			connect.WithClientOptions(opts...),
-		),
-		leaveFocusByTarget: connect.NewClient[v1.PluginHostServiceLeaveFocusByTargetRequest, v1.PluginHostServiceLeaveFocusByTargetResponse](
-			httpClient,
-			baseURL+PluginHostServiceLeaveFocusByTargetProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("LeaveFocusByTarget")),
-			connect.WithClientOptions(opts...),
-		),
-		presentFocus: connect.NewClient[v1.PluginHostServicePresentFocusRequest, v1.PluginHostServicePresentFocusResponse](
-			httpClient,
-			baseURL+PluginHostServicePresentFocusProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("PresentFocus")),
-			connect.WithClientOptions(opts...),
-		),
-		queryStreamHistory: connect.NewClient[v1.PluginHostServiceQueryStreamHistoryRequest, v1.PluginHostServiceQueryStreamHistoryResponse](
-			httpClient,
-			baseURL+PluginHostServiceQueryStreamHistoryProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("QueryStreamHistory")),
-			connect.WithClientOptions(opts...),
-		),
-		decryptOwnAuditRows: connect.NewClient[v1.DecryptOwnAuditRowsRequest, v1.DecryptOwnAuditRowsResponse](
-			httpClient,
-			baseURL+PluginHostServiceDecryptOwnAuditRowsProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("DecryptOwnAuditRows")),
-			connect.WithClientOptions(opts...),
-		),
-		requestEmitToken: connect.NewClient[v1.PluginHostServiceRequestEmitTokenRequest, v1.PluginHostServiceRequestEmitTokenResponse](
-			httpClient,
-			baseURL+PluginHostServiceRequestEmitTokenProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("RequestEmitToken")),
-			connect.WithClientOptions(opts...),
-		),
-		setConnectionFocus: connect.NewClient[v1.PluginHostServiceSetConnectionFocusRequest, v1.PluginHostServiceSetConnectionFocusResponse](
-			httpClient,
-			baseURL+PluginHostServiceSetConnectionFocusProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("SetConnectionFocus")),
-			connect.WithClientOptions(opts...),
-		),
-		getConnectionFocus: connect.NewClient[v1.PluginHostServiceGetConnectionFocusRequest, v1.PluginHostServiceGetConnectionFocusResponse](
-			httpClient,
-			baseURL+PluginHostServiceGetConnectionFocusProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("GetConnectionFocus")),
-			connect.WithClientOptions(opts...),
-		),
-		autoFocusOnJoin: connect.NewClient[v1.PluginHostServiceAutoFocusOnJoinRequest, v1.PluginHostServiceAutoFocusOnJoinResponse](
-			httpClient,
-			baseURL+PluginHostServiceAutoFocusOnJoinProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("AutoFocusOnJoin")),
-			connect.WithClientOptions(opts...),
-		),
-		isAnyConnFocused: connect.NewClient[v1.PluginHostServiceIsAnyConnFocusedRequest, v1.PluginHostServiceIsAnyConnFocusedResponse](
-			httpClient,
-			baseURL+PluginHostServiceIsAnyConnFocusedProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("IsAnyConnFocused")),
-			connect.WithClientOptions(opts...),
-		),
-		evaluate: connect.NewClient[v1.PluginHostServiceEvaluateRequest, v1.PluginHostServiceEvaluateResponse](
-			httpClient,
-			baseURL+PluginHostServiceEvaluateProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("Evaluate")),
-			connect.WithClientOptions(opts...),
-		),
-		listCommands: connect.NewClient[v1.PluginHostServiceListCommandsRequest, v1.PluginHostServiceListCommandsResponse](
-			httpClient,
-			baseURL+PluginHostServiceListCommandsProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("ListCommands")),
-			connect.WithClientOptions(opts...),
-		),
-		getCommandHelp: connect.NewClient[v1.PluginHostServiceGetCommandHelpRequest, v1.PluginHostServiceGetCommandHelpResponse](
-			httpClient,
-			baseURL+PluginHostServiceGetCommandHelpProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("GetCommandHelp")),
-			connect.WithClientOptions(opts...),
-		),
-		getSetting: connect.NewClient[v1.PluginHostServiceGetSettingRequest, v1.PluginHostServiceGetSettingResponse](
-			httpClient,
-			baseURL+PluginHostServiceGetSettingProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("GetSetting")),
-			connect.WithClientOptions(opts...),
-		),
-		setSetting: connect.NewClient[v1.PluginHostServiceSetSettingRequest, v1.PluginHostServiceSetSettingResponse](
-			httpClient,
-			baseURL+PluginHostServiceSetSettingProcedure,
-			connect.WithSchema(pluginHostServiceMethods.ByName("SetSetting")),
-			connect.WithClientOptions(opts...),
-		),
-	}
-}
-
-// pluginHostServiceClient implements PluginHostServiceClient.
-type pluginHostServiceClient struct {
-	emitEvent           *connect.Client[v1.PluginHostServiceEmitEventRequest, v1.PluginHostServiceEmitEventResponse]
-	log                 *connect.Client[v1.PluginHostServiceLogRequest, v1.PluginHostServiceLogResponse]
-	kVGet               *connect.Client[v1.PluginHostServiceKVGetRequest, v1.PluginHostServiceKVGetResponse]
-	kVSet               *connect.Client[v1.PluginHostServiceKVSetRequest, v1.PluginHostServiceKVSetResponse]
-	kVDelete            *connect.Client[v1.PluginHostServiceKVDeleteRequest, v1.PluginHostServiceKVDeleteResponse]
-	addSessionStream    *connect.Client[v1.PluginHostServiceAddSessionStreamRequest, v1.PluginHostServiceAddSessionStreamResponse]
-	removeSessionStream *connect.Client[v1.PluginHostServiceRemoveSessionStreamRequest, v1.PluginHostServiceRemoveSessionStreamResponse]
-	joinFocus           *connect.Client[v1.PluginHostServiceJoinFocusRequest, v1.PluginHostServiceJoinFocusResponse]
-	leaveFocus          *connect.Client[v1.PluginHostServiceLeaveFocusRequest, v1.PluginHostServiceLeaveFocusResponse]
-	leaveFocusByTarget  *connect.Client[v1.PluginHostServiceLeaveFocusByTargetRequest, v1.PluginHostServiceLeaveFocusByTargetResponse]
-	presentFocus        *connect.Client[v1.PluginHostServicePresentFocusRequest, v1.PluginHostServicePresentFocusResponse]
-	queryStreamHistory  *connect.Client[v1.PluginHostServiceQueryStreamHistoryRequest, v1.PluginHostServiceQueryStreamHistoryResponse]
-	decryptOwnAuditRows *connect.Client[v1.DecryptOwnAuditRowsRequest, v1.DecryptOwnAuditRowsResponse]
-	requestEmitToken    *connect.Client[v1.PluginHostServiceRequestEmitTokenRequest, v1.PluginHostServiceRequestEmitTokenResponse]
-	setConnectionFocus  *connect.Client[v1.PluginHostServiceSetConnectionFocusRequest, v1.PluginHostServiceSetConnectionFocusResponse]
-	getConnectionFocus  *connect.Client[v1.PluginHostServiceGetConnectionFocusRequest, v1.PluginHostServiceGetConnectionFocusResponse]
-	autoFocusOnJoin     *connect.Client[v1.PluginHostServiceAutoFocusOnJoinRequest, v1.PluginHostServiceAutoFocusOnJoinResponse]
-	isAnyConnFocused    *connect.Client[v1.PluginHostServiceIsAnyConnFocusedRequest, v1.PluginHostServiceIsAnyConnFocusedResponse]
-	evaluate            *connect.Client[v1.PluginHostServiceEvaluateRequest, v1.PluginHostServiceEvaluateResponse]
-	listCommands        *connect.Client[v1.PluginHostServiceListCommandsRequest, v1.PluginHostServiceListCommandsResponse]
-	getCommandHelp      *connect.Client[v1.PluginHostServiceGetCommandHelpRequest, v1.PluginHostServiceGetCommandHelpResponse]
-	getSetting          *connect.Client[v1.PluginHostServiceGetSettingRequest, v1.PluginHostServiceGetSettingResponse]
-	setSetting          *connect.Client[v1.PluginHostServiceSetSettingRequest, v1.PluginHostServiceSetSettingResponse]
-}
-
-// EmitEvent calls holomush.plugin.v1.PluginHostService.EmitEvent.
-func (c *pluginHostServiceClient) EmitEvent(ctx context.Context, req *connect.Request[v1.PluginHostServiceEmitEventRequest]) (*connect.Response[v1.PluginHostServiceEmitEventResponse], error) {
-	return c.emitEvent.CallUnary(ctx, req)
-}
-
-// Log calls holomush.plugin.v1.PluginHostService.Log.
-func (c *pluginHostServiceClient) Log(ctx context.Context, req *connect.Request[v1.PluginHostServiceLogRequest]) (*connect.Response[v1.PluginHostServiceLogResponse], error) {
-	return c.log.CallUnary(ctx, req)
-}
-
-// KVGet calls holomush.plugin.v1.PluginHostService.KVGet.
-func (c *pluginHostServiceClient) KVGet(ctx context.Context, req *connect.Request[v1.PluginHostServiceKVGetRequest]) (*connect.Response[v1.PluginHostServiceKVGetResponse], error) {
-	return c.kVGet.CallUnary(ctx, req)
-}
-
-// KVSet calls holomush.plugin.v1.PluginHostService.KVSet.
-func (c *pluginHostServiceClient) KVSet(ctx context.Context, req *connect.Request[v1.PluginHostServiceKVSetRequest]) (*connect.Response[v1.PluginHostServiceKVSetResponse], error) {
-	return c.kVSet.CallUnary(ctx, req)
-}
-
-// KVDelete calls holomush.plugin.v1.PluginHostService.KVDelete.
-func (c *pluginHostServiceClient) KVDelete(ctx context.Context, req *connect.Request[v1.PluginHostServiceKVDeleteRequest]) (*connect.Response[v1.PluginHostServiceKVDeleteResponse], error) {
-	return c.kVDelete.CallUnary(ctx, req)
-}
-
-// AddSessionStream calls holomush.plugin.v1.PluginHostService.AddSessionStream.
-func (c *pluginHostServiceClient) AddSessionStream(ctx context.Context, req *connect.Request[v1.PluginHostServiceAddSessionStreamRequest]) (*connect.Response[v1.PluginHostServiceAddSessionStreamResponse], error) {
-	return c.addSessionStream.CallUnary(ctx, req)
-}
-
-// RemoveSessionStream calls holomush.plugin.v1.PluginHostService.RemoveSessionStream.
-func (c *pluginHostServiceClient) RemoveSessionStream(ctx context.Context, req *connect.Request[v1.PluginHostServiceRemoveSessionStreamRequest]) (*connect.Response[v1.PluginHostServiceRemoveSessionStreamResponse], error) {
-	return c.removeSessionStream.CallUnary(ctx, req)
-}
-
-// JoinFocus calls holomush.plugin.v1.PluginHostService.JoinFocus.
-func (c *pluginHostServiceClient) JoinFocus(ctx context.Context, req *connect.Request[v1.PluginHostServiceJoinFocusRequest]) (*connect.Response[v1.PluginHostServiceJoinFocusResponse], error) {
-	return c.joinFocus.CallUnary(ctx, req)
-}
-
-// LeaveFocus calls holomush.plugin.v1.PluginHostService.LeaveFocus.
-func (c *pluginHostServiceClient) LeaveFocus(ctx context.Context, req *connect.Request[v1.PluginHostServiceLeaveFocusRequest]) (*connect.Response[v1.PluginHostServiceLeaveFocusResponse], error) {
-	return c.leaveFocus.CallUnary(ctx, req)
-}
-
-// LeaveFocusByTarget calls holomush.plugin.v1.PluginHostService.LeaveFocusByTarget.
-func (c *pluginHostServiceClient) LeaveFocusByTarget(ctx context.Context, req *connect.Request[v1.PluginHostServiceLeaveFocusByTargetRequest]) (*connect.Response[v1.PluginHostServiceLeaveFocusByTargetResponse], error) {
-	return c.leaveFocusByTarget.CallUnary(ctx, req)
-}
-
-// PresentFocus calls holomush.plugin.v1.PluginHostService.PresentFocus.
-func (c *pluginHostServiceClient) PresentFocus(ctx context.Context, req *connect.Request[v1.PluginHostServicePresentFocusRequest]) (*connect.Response[v1.PluginHostServicePresentFocusResponse], error) {
-	return c.presentFocus.CallUnary(ctx, req)
-}
-
-// QueryStreamHistory calls holomush.plugin.v1.PluginHostService.QueryStreamHistory.
-func (c *pluginHostServiceClient) QueryStreamHistory(ctx context.Context, req *connect.Request[v1.PluginHostServiceQueryStreamHistoryRequest]) (*connect.Response[v1.PluginHostServiceQueryStreamHistoryResponse], error) {
-	return c.queryStreamHistory.CallUnary(ctx, req)
-}
-
-// DecryptOwnAuditRows calls holomush.plugin.v1.PluginHostService.DecryptOwnAuditRows.
-func (c *pluginHostServiceClient) DecryptOwnAuditRows(ctx context.Context, req *connect.Request[v1.DecryptOwnAuditRowsRequest]) (*connect.Response[v1.DecryptOwnAuditRowsResponse], error) {
-	return c.decryptOwnAuditRows.CallUnary(ctx, req)
-}
-
-// RequestEmitToken calls holomush.plugin.v1.PluginHostService.RequestEmitToken.
-func (c *pluginHostServiceClient) RequestEmitToken(ctx context.Context, req *connect.Request[v1.PluginHostServiceRequestEmitTokenRequest]) (*connect.Response[v1.PluginHostServiceRequestEmitTokenResponse], error) {
-	return c.requestEmitToken.CallUnary(ctx, req)
-}
-
-// SetConnectionFocus calls holomush.plugin.v1.PluginHostService.SetConnectionFocus.
-func (c *pluginHostServiceClient) SetConnectionFocus(ctx context.Context, req *connect.Request[v1.PluginHostServiceSetConnectionFocusRequest]) (*connect.Response[v1.PluginHostServiceSetConnectionFocusResponse], error) {
-	return c.setConnectionFocus.CallUnary(ctx, req)
-}
-
-// GetConnectionFocus calls holomush.plugin.v1.PluginHostService.GetConnectionFocus.
-func (c *pluginHostServiceClient) GetConnectionFocus(ctx context.Context, req *connect.Request[v1.PluginHostServiceGetConnectionFocusRequest]) (*connect.Response[v1.PluginHostServiceGetConnectionFocusResponse], error) {
-	return c.getConnectionFocus.CallUnary(ctx, req)
-}
-
-// AutoFocusOnJoin calls holomush.plugin.v1.PluginHostService.AutoFocusOnJoin.
-func (c *pluginHostServiceClient) AutoFocusOnJoin(ctx context.Context, req *connect.Request[v1.PluginHostServiceAutoFocusOnJoinRequest]) (*connect.Response[v1.PluginHostServiceAutoFocusOnJoinResponse], error) {
-	return c.autoFocusOnJoin.CallUnary(ctx, req)
-}
-
-// IsAnyConnFocused calls holomush.plugin.v1.PluginHostService.IsAnyConnFocused.
-func (c *pluginHostServiceClient) IsAnyConnFocused(ctx context.Context, req *connect.Request[v1.PluginHostServiceIsAnyConnFocusedRequest]) (*connect.Response[v1.PluginHostServiceIsAnyConnFocusedResponse], error) {
-	return c.isAnyConnFocused.CallUnary(ctx, req)
-}
-
-// Evaluate calls holomush.plugin.v1.PluginHostService.Evaluate.
-func (c *pluginHostServiceClient) Evaluate(ctx context.Context, req *connect.Request[v1.PluginHostServiceEvaluateRequest]) (*connect.Response[v1.PluginHostServiceEvaluateResponse], error) {
-	return c.evaluate.CallUnary(ctx, req)
-}
-
-// ListCommands calls holomush.plugin.v1.PluginHostService.ListCommands.
-func (c *pluginHostServiceClient) ListCommands(ctx context.Context, req *connect.Request[v1.PluginHostServiceListCommandsRequest]) (*connect.Response[v1.PluginHostServiceListCommandsResponse], error) {
-	return c.listCommands.CallUnary(ctx, req)
-}
-
-// GetCommandHelp calls holomush.plugin.v1.PluginHostService.GetCommandHelp.
-func (c *pluginHostServiceClient) GetCommandHelp(ctx context.Context, req *connect.Request[v1.PluginHostServiceGetCommandHelpRequest]) (*connect.Response[v1.PluginHostServiceGetCommandHelpResponse], error) {
-	return c.getCommandHelp.CallUnary(ctx, req)
-}
-
-// GetSetting calls holomush.plugin.v1.PluginHostService.GetSetting.
-func (c *pluginHostServiceClient) GetSetting(ctx context.Context, req *connect.Request[v1.PluginHostServiceGetSettingRequest]) (*connect.Response[v1.PluginHostServiceGetSettingResponse], error) {
-	return c.getSetting.CallUnary(ctx, req)
-}
-
-// SetSetting calls holomush.plugin.v1.PluginHostService.SetSetting.
-func (c *pluginHostServiceClient) SetSetting(ctx context.Context, req *connect.Request[v1.PluginHostServiceSetSettingRequest]) (*connect.Response[v1.PluginHostServiceSetSettingResponse], error) {
-	return c.setSetting.CallUnary(ctx, req)
-}
-
-// PluginHostServiceHandler is an implementation of the holomush.plugin.v1.PluginHostService
-// service.
-type PluginHostServiceHandler interface {
-	// EmitEvent publishes one plugin-originated event onto the bus through the
-	// host emit fence. SERVED: pluginHostServiceServer.EmitEvent. The caller's
-	// identity is NOT trusted from the wire — the plugin presents a host-issued
-	// dispatch token in the x-holomush-emit-token metadata header, the host
-	// recovers the vouched-for actor from tokenStore.Lookup(pluginName, token),
-	// and a missing/foreign token is rejected (EMIT_TOKEN_MISSING /
-	// EMIT_TOKEN_REJECTED). The recovered actor then flows through
-	// PluginEventEmitter.Emit, which enforces the manifest gates: emits (subject
-	// namespace must be declared), actor_kinds_claimable (actor kind must be
-	// listed — EMIT_ACTOR_KIND_NOT_CLAIMABLE), and the crypto.emits sensitivity
-	// fence. These gates fire identically for Lua and binary plugins (plugin
-	// runtime symmetry); the token mechanism is the binary-side forgery fence.
-	EmitEvent(context.Context, *connect.Request[v1.PluginHostServiceEmitEventRequest]) (*connect.Response[v1.PluginHostServiceEmitEventResponse], error)
-	// Log forwards a plugin log line to the host logger. DECLARED BUT UNSERVED:
-	// pluginHostServiceServer does not implement it and no production client
-	// calls it, so it currently returns codes.Unimplemented (holomush-l6std).
-	Log(context.Context, *connect.Request[v1.PluginHostServiceLogRequest]) (*connect.Response[v1.PluginHostServiceLogResponse], error)
-	// KVGet reads a value from the plugin's namespaced key-value store.
-	// DECLARED BUT UNSERVED (holomush-l6std): no server impl, no client; returns
-	// codes.Unimplemented today.
-	KVGet(context.Context, *connect.Request[v1.PluginHostServiceKVGetRequest]) (*connect.Response[v1.PluginHostServiceKVGetResponse], error)
-	// KVSet writes a value into the plugin's namespaced key-value store.
-	// DECLARED BUT UNSERVED (holomush-l6std): no server impl, no client; returns
-	// codes.Unimplemented today.
-	KVSet(context.Context, *connect.Request[v1.PluginHostServiceKVSetRequest]) (*connect.Response[v1.PluginHostServiceKVSetResponse], error)
-	// KVDelete removes a key from the plugin's namespaced key-value store.
-	// DECLARED BUT UNSERVED (holomush-l6std): no server impl, no client; returns
-	// codes.Unimplemented today.
-	KVDelete(context.Context, *connect.Request[v1.PluginHostServiceKVDeleteRequest]) (*connect.Response[v1.PluginHostServiceKVDeleteResponse], error)
-	// AddSessionStream subscribes an active session to an additional stream
-	// mid-session. DECLARED BUT UNSERVED (holomush-l6std): no server impl, no
-	// production client; returns codes.Unimplemented today. (The wire comment
-	// about SESSION_NOT_FOUND describes the intended-but-unimplemented contract.)
-	AddSessionStream(context.Context, *connect.Request[v1.PluginHostServiceAddSessionStreamRequest]) (*connect.Response[v1.PluginHostServiceAddSessionStreamResponse], error)
-	// RemoveSessionStream unsubscribes an active session from a stream.
-	// DECLARED BUT UNSERVED (holomush-l6std): no server impl, no production
-	// client; returns codes.Unimplemented today.
-	RemoveSessionStream(context.Context, *connect.Request[v1.PluginHostServiceRemoveSessionStreamRequest]) (*connect.Response[v1.PluginHostServiceRemoveSessionStreamResponse], error)
-	// JoinFocus adds a focus membership (e.g. a scene) to a session via the
-	// host focus coordinator. SERVED: pluginHostServiceServer.JoinFocus. The
-	// plugin declares intent; the coordinator applies the kind-specific replay
-	// policy. Fails if the focus coordinator is not configured.
-	JoinFocus(context.Context, *connect.Request[v1.PluginHostServiceJoinFocusRequest]) (*connect.Response[v1.PluginHostServiceJoinFocusResponse], error)
-	// LeaveFocus removes one focus membership from a session. SERVED:
-	// pluginHostServiceServer.LeaveFocus. Idempotent — leaving a target the
-	// session does not hold is a successful no-op.
-	LeaveFocus(context.Context, *connect.Request[v1.PluginHostServiceLeaveFocusRequest]) (*connect.Response[v1.PluginHostServiceLeaveFocusResponse], error)
-	// LeaveFocusByTarget removes the given focus membership from every
-	// non-expired session that holds it — cross-session fan-out (e.g. a
-	// scene-end reaching all participants). SERVED:
-	// pluginHostServiceServer.LeaveFocusByTarget. Partial success is normal:
-	// per-session failures are aggregated into the response (succeeded /
-	// total_scanned / failed_session_ids) rather than aborting the sweep; the
-	// RPC error is reserved for the enumeration step itself failing.
-	LeaveFocusByTarget(context.Context, *connect.Request[v1.PluginHostServiceLeaveFocusByTargetRequest]) (*connect.Response[v1.PluginHostServiceLeaveFocusByTargetResponse], error)
-	// PresentFocus repoints a session's PresentingFocus to an existing
-	// membership. SERVED: pluginHostServiceServer.PresentFocus. The target MUST
-	// already be in the session's FocusMemberships (membership is validated, not
-	// implicitly created).
-	PresentFocus(context.Context, *connect.Request[v1.PluginHostServicePresentFocusRequest]) (*connect.Response[v1.PluginHostServicePresentFocusResponse], error)
-	// QueryStreamHistory reads the tail of a stream for plugin-side display,
-	// backward-paginated by opaque cursor. SERVED:
-	// pluginHostServiceServer.QueryStreamHistory via HistoryReader.ReplayTail.
-	// Read-only: it does not advance session cursors or mutate session state.
-	// A negative count is rejected (INVALID_ARGUMENT); count is CLAMPED to
-	// maxQueryStreamHistoryCount (500), not rejected, when too large.
-	QueryStreamHistory(context.Context, *connect.Request[v1.PluginHostServiceQueryStreamHistoryRequest]) (*connect.Response[v1.PluginHostServiceQueryStreamHistoryResponse], error)
-	// DecryptOwnAuditRows decrypts a batch of the calling plugin's OWN encrypted
-	// audit rows host-side; the plugin never holds a DEK. SERVED:
-	// pluginHostServiceServer.DecryptOwnAuditRows via ReadbackDecryptor.
-	// DecryptOwnRows. Authorization is two-gate (INV-CRYPTO-27): OwnerMap subject
-	// ownership (g1) plus the crypto.emits[].readback manifest flag (g2). Each
-	// input row gets an independent RowResult (INV-CRYPTO-37) carrying either
-	// plaintext or a stable snake_case no_plaintext_reason ("not_owner",
-	// "auth_guard_deny", "dek_missing", "downgrade_refused", "stale_dek",
-	// "audit_queue_full", "internal"; readback.go reasonToWire). Request /
-	// response shapes and RowResult live in audit.proto (AuditRow domain).
-	DecryptOwnAuditRows(context.Context, *connect.Request[v1.DecryptOwnAuditRowsRequest]) (*connect.Response[v1.DecryptOwnAuditRowsResponse], error)
-	// RequestEmitToken issues a self-token bound to {ActorPlugin, pluginName} so
-	// a plugin-served gRPC handler (e.g. SceneService.CreateScene) — which is NOT
-	// reached via DeliverEvent/DeliverCommand and so holds no dispatch token —
-	// can still call EmitEvent. SERVED: pluginHostServiceServer.RequestEmitToken.
-	// The plugin's identity is taken from the mTLS-bound server struct
-	// (s.pluginName); the request carries no identity fields, so a plugin cannot
-	// impersonate another actor or escalate to a character actor through this
-	// RPC. The actor_kinds_claimable manifest gate still fires when the issued
-	// token is later spent at EmitEvent. (Spec §3.3.5 / §5.4 two-token pattern.)
-	RequestEmitToken(context.Context, *connect.Request[v1.PluginHostServiceRequestEmitTokenRequest]) (*connect.Response[v1.PluginHostServiceRequestEmitTokenResponse], error)
-	// SetConnectionFocus is the Phase-5 explicit focus mutation for a single
-	// Connection. SERVED: pluginHostServiceServer.SetConnectionFocus. The
-	// substrate validates the requested membership against the session's
-	// FocusMemberships (D4), then writes Connection.FocusKey and (D9-gated)
-	// Info.PresentingFocus atomically under one Store-lock acquisition (D7).
-	SetConnectionFocus(context.Context, *connect.Request[v1.PluginHostServiceSetConnectionFocusRequest]) (*connect.Response[v1.PluginHostServiceSetConnectionFocusResponse], error)
-	// GetConnectionFocus returns the named connection's current per-connection
-	// focus, or absent when the connection is grid-focused (FocusKey nil) or
-	// unknown. Read-only counterpart of SetConnectionFocus; lets plugins route
-	// connection-scoped operations (e.g. scene pose) to the focused target.
-	// See goplugin/host_service.go::GetConnectionFocus.
-	GetConnectionFocus(context.Context, *connect.Request[v1.PluginHostServiceGetConnectionFocusRequest]) (*connect.Response[v1.PluginHostServiceGetConnectionFocusResponse], error)
-	// AutoFocusOnJoin is the Phase-5 fan-out that focuses all of a character's
-	// terminal/telnet connections on a scene at once. SERVED:
-	// pluginHostServiceServer.AutoFocusOnJoin. Connections already explicitly
-	// focused elsewhere are skipped (D8). The caller MUST have completed
-	// JoinFocus first, since the substrate requires the membership to exist.
-	AutoFocusOnJoin(context.Context, *connect.Request[v1.PluginHostServiceAutoFocusOnJoinRequest]) (*connect.Response[v1.PluginHostServiceAutoFocusOnJoinResponse], error)
-	// IsAnyConnFocused is the Phase-5 notification-emission helper: it reports
-	// whether any of the character's connections currently focuses the given
-	// scene, so callers can decide whether to emit a focus-related notification.
-	// SERVED: pluginHostServiceServer.IsAnyConnFocused.
-	IsAnyConnFocused(context.Context, *connect.Request[v1.PluginHostServiceIsAnyConnFocusedRequest]) (*connect.Response[v1.PluginHostServiceIsAnyConnFocusedResponse], error)
-	// Evaluate runs the host ABAC engine for one action against one resource
-	// instance owned by the calling plugin. SERVED:
-	// pluginHostServiceServer.Evaluate. The subject is derived host-side from the
-	// dispatch token exactly as EmitEvent does (token→actor recovery) — there is
-	// no subject field on the wire (spec §2, INV-PLUGIN-22). Fails closed on nil engine,
-	// missing/rejected token, empty actor subject, or a resource type the plugin
-	// does not own.
-	Evaluate(context.Context, *connect.Request[v1.PluginHostServiceEvaluateRequest]) (*connect.Response[v1.PluginHostServiceEvaluateResponse], error)
-	// ListCommands enumerates the commands the named character may execute,
-	// ABAC-filtered by the host. SERVED: pluginHostServiceServer.ListCommands,
-	// delegating to commandquery.Querier.Available. The subject is the request's
-	// character_id (parity with the Lua holomush.list_commands(character_id) host
-	// function — not the dispatch-token actor, since this is read-only metadata,
-	// not an actor-gated mutation). incomplete is true when engine errors hid
-	// some commands.
-	ListCommands(context.Context, *connect.Request[v1.PluginHostServiceListCommandsRequest]) (*connect.Response[v1.PluginHostServiceListCommandsResponse], error)
-	// GetCommandHelp returns full help detail for one command after an access
-	// check for character_id. SERVED: pluginHostServiceServer.GetCommandHelp,
-	// delegating to commandquery.Querier.Help. Mirrors the Lua
-	// holomush.get_command_help(name, character_id) host function.
-	GetCommandHelp(context.Context, *connect.Request[v1.PluginHostServiceGetCommandHelpRequest]) (*connect.Response[v1.PluginHostServiceGetCommandHelpResponse], error)
-	// GetSetting reads a single-scope setting in the calling plugin's owner
-	// partition (owner bound host-side from the authenticated plugin name, never
-	// from the request). The handler resolves SettingScope to its backing store;
-	// a missing key returns a successful response with found=false, never a
-	// codes.NotFound status error.
-	GetSetting(context.Context, *connect.Request[v1.PluginHostServiceGetSettingRequest]) (*connect.Response[v1.PluginHostServiceGetSettingResponse], error)
-	// SetSetting writes a single-scope setting in the calling plugin's partition;
-	// GAME scope requires an operator authorization decision (host-enforced, not
-	// trusted from the wire). The owner partition is bound host-side from the
-	// authenticated plugin name.
-	SetSetting(context.Context, *connect.Request[v1.PluginHostServiceSetSettingRequest]) (*connect.Response[v1.PluginHostServiceSetSettingResponse], error)
-}
-
-// NewPluginHostServiceHandler builds an HTTP handler from the service implementation. It returns
-// the path on which to mount the handler and the handler itself.
-//
-// By default, handlers support the Connect, gRPC, and gRPC-Web protocols with the binary Protobuf
-// and JSON codecs. They also support gzip compression.
-func NewPluginHostServiceHandler(svc PluginHostServiceHandler, opts ...connect.HandlerOption) (string, http.Handler) {
-	pluginHostServiceMethods := v1.File_holomush_plugin_v1_plugin_proto.Services().ByName("PluginHostService").Methods()
-	pluginHostServiceEmitEventHandler := connect.NewUnaryHandler(
-		PluginHostServiceEmitEventProcedure,
-		svc.EmitEvent,
-		connect.WithSchema(pluginHostServiceMethods.ByName("EmitEvent")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceLogHandler := connect.NewUnaryHandler(
-		PluginHostServiceLogProcedure,
-		svc.Log,
-		connect.WithSchema(pluginHostServiceMethods.ByName("Log")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceKVGetHandler := connect.NewUnaryHandler(
-		PluginHostServiceKVGetProcedure,
-		svc.KVGet,
-		connect.WithSchema(pluginHostServiceMethods.ByName("KVGet")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceKVSetHandler := connect.NewUnaryHandler(
-		PluginHostServiceKVSetProcedure,
-		svc.KVSet,
-		connect.WithSchema(pluginHostServiceMethods.ByName("KVSet")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceKVDeleteHandler := connect.NewUnaryHandler(
-		PluginHostServiceKVDeleteProcedure,
-		svc.KVDelete,
-		connect.WithSchema(pluginHostServiceMethods.ByName("KVDelete")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceAddSessionStreamHandler := connect.NewUnaryHandler(
-		PluginHostServiceAddSessionStreamProcedure,
-		svc.AddSessionStream,
-		connect.WithSchema(pluginHostServiceMethods.ByName("AddSessionStream")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceRemoveSessionStreamHandler := connect.NewUnaryHandler(
-		PluginHostServiceRemoveSessionStreamProcedure,
-		svc.RemoveSessionStream,
-		connect.WithSchema(pluginHostServiceMethods.ByName("RemoveSessionStream")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceJoinFocusHandler := connect.NewUnaryHandler(
-		PluginHostServiceJoinFocusProcedure,
-		svc.JoinFocus,
-		connect.WithSchema(pluginHostServiceMethods.ByName("JoinFocus")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceLeaveFocusHandler := connect.NewUnaryHandler(
-		PluginHostServiceLeaveFocusProcedure,
-		svc.LeaveFocus,
-		connect.WithSchema(pluginHostServiceMethods.ByName("LeaveFocus")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceLeaveFocusByTargetHandler := connect.NewUnaryHandler(
-		PluginHostServiceLeaveFocusByTargetProcedure,
-		svc.LeaveFocusByTarget,
-		connect.WithSchema(pluginHostServiceMethods.ByName("LeaveFocusByTarget")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServicePresentFocusHandler := connect.NewUnaryHandler(
-		PluginHostServicePresentFocusProcedure,
-		svc.PresentFocus,
-		connect.WithSchema(pluginHostServiceMethods.ByName("PresentFocus")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceQueryStreamHistoryHandler := connect.NewUnaryHandler(
-		PluginHostServiceQueryStreamHistoryProcedure,
-		svc.QueryStreamHistory,
-		connect.WithSchema(pluginHostServiceMethods.ByName("QueryStreamHistory")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceDecryptOwnAuditRowsHandler := connect.NewUnaryHandler(
-		PluginHostServiceDecryptOwnAuditRowsProcedure,
-		svc.DecryptOwnAuditRows,
-		connect.WithSchema(pluginHostServiceMethods.ByName("DecryptOwnAuditRows")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceRequestEmitTokenHandler := connect.NewUnaryHandler(
-		PluginHostServiceRequestEmitTokenProcedure,
-		svc.RequestEmitToken,
-		connect.WithSchema(pluginHostServiceMethods.ByName("RequestEmitToken")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceSetConnectionFocusHandler := connect.NewUnaryHandler(
-		PluginHostServiceSetConnectionFocusProcedure,
-		svc.SetConnectionFocus,
-		connect.WithSchema(pluginHostServiceMethods.ByName("SetConnectionFocus")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceGetConnectionFocusHandler := connect.NewUnaryHandler(
-		PluginHostServiceGetConnectionFocusProcedure,
-		svc.GetConnectionFocus,
-		connect.WithSchema(pluginHostServiceMethods.ByName("GetConnectionFocus")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceAutoFocusOnJoinHandler := connect.NewUnaryHandler(
-		PluginHostServiceAutoFocusOnJoinProcedure,
-		svc.AutoFocusOnJoin,
-		connect.WithSchema(pluginHostServiceMethods.ByName("AutoFocusOnJoin")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceIsAnyConnFocusedHandler := connect.NewUnaryHandler(
-		PluginHostServiceIsAnyConnFocusedProcedure,
-		svc.IsAnyConnFocused,
-		connect.WithSchema(pluginHostServiceMethods.ByName("IsAnyConnFocused")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceEvaluateHandler := connect.NewUnaryHandler(
-		PluginHostServiceEvaluateProcedure,
-		svc.Evaluate,
-		connect.WithSchema(pluginHostServiceMethods.ByName("Evaluate")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceListCommandsHandler := connect.NewUnaryHandler(
-		PluginHostServiceListCommandsProcedure,
-		svc.ListCommands,
-		connect.WithSchema(pluginHostServiceMethods.ByName("ListCommands")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceGetCommandHelpHandler := connect.NewUnaryHandler(
-		PluginHostServiceGetCommandHelpProcedure,
-		svc.GetCommandHelp,
-		connect.WithSchema(pluginHostServiceMethods.ByName("GetCommandHelp")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceGetSettingHandler := connect.NewUnaryHandler(
-		PluginHostServiceGetSettingProcedure,
-		svc.GetSetting,
-		connect.WithSchema(pluginHostServiceMethods.ByName("GetSetting")),
-		connect.WithHandlerOptions(opts...),
-	)
-	pluginHostServiceSetSettingHandler := connect.NewUnaryHandler(
-		PluginHostServiceSetSettingProcedure,
-		svc.SetSetting,
-		connect.WithSchema(pluginHostServiceMethods.ByName("SetSetting")),
-		connect.WithHandlerOptions(opts...),
-	)
-	return "/holomush.plugin.v1.PluginHostService/", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		switch r.URL.Path {
-		case PluginHostServiceEmitEventProcedure:
-			pluginHostServiceEmitEventHandler.ServeHTTP(w, r)
-		case PluginHostServiceLogProcedure:
-			pluginHostServiceLogHandler.ServeHTTP(w, r)
-		case PluginHostServiceKVGetProcedure:
-			pluginHostServiceKVGetHandler.ServeHTTP(w, r)
-		case PluginHostServiceKVSetProcedure:
-			pluginHostServiceKVSetHandler.ServeHTTP(w, r)
-		case PluginHostServiceKVDeleteProcedure:
-			pluginHostServiceKVDeleteHandler.ServeHTTP(w, r)
-		case PluginHostServiceAddSessionStreamProcedure:
-			pluginHostServiceAddSessionStreamHandler.ServeHTTP(w, r)
-		case PluginHostServiceRemoveSessionStreamProcedure:
-			pluginHostServiceRemoveSessionStreamHandler.ServeHTTP(w, r)
-		case PluginHostServiceJoinFocusProcedure:
-			pluginHostServiceJoinFocusHandler.ServeHTTP(w, r)
-		case PluginHostServiceLeaveFocusProcedure:
-			pluginHostServiceLeaveFocusHandler.ServeHTTP(w, r)
-		case PluginHostServiceLeaveFocusByTargetProcedure:
-			pluginHostServiceLeaveFocusByTargetHandler.ServeHTTP(w, r)
-		case PluginHostServicePresentFocusProcedure:
-			pluginHostServicePresentFocusHandler.ServeHTTP(w, r)
-		case PluginHostServiceQueryStreamHistoryProcedure:
-			pluginHostServiceQueryStreamHistoryHandler.ServeHTTP(w, r)
-		case PluginHostServiceDecryptOwnAuditRowsProcedure:
-			pluginHostServiceDecryptOwnAuditRowsHandler.ServeHTTP(w, r)
-		case PluginHostServiceRequestEmitTokenProcedure:
-			pluginHostServiceRequestEmitTokenHandler.ServeHTTP(w, r)
-		case PluginHostServiceSetConnectionFocusProcedure:
-			pluginHostServiceSetConnectionFocusHandler.ServeHTTP(w, r)
-		case PluginHostServiceGetConnectionFocusProcedure:
-			pluginHostServiceGetConnectionFocusHandler.ServeHTTP(w, r)
-		case PluginHostServiceAutoFocusOnJoinProcedure:
-			pluginHostServiceAutoFocusOnJoinHandler.ServeHTTP(w, r)
-		case PluginHostServiceIsAnyConnFocusedProcedure:
-			pluginHostServiceIsAnyConnFocusedHandler.ServeHTTP(w, r)
-		case PluginHostServiceEvaluateProcedure:
-			pluginHostServiceEvaluateHandler.ServeHTTP(w, r)
-		case PluginHostServiceListCommandsProcedure:
-			pluginHostServiceListCommandsHandler.ServeHTTP(w, r)
-		case PluginHostServiceGetCommandHelpProcedure:
-			pluginHostServiceGetCommandHelpHandler.ServeHTTP(w, r)
-		case PluginHostServiceGetSettingProcedure:
-			pluginHostServiceGetSettingHandler.ServeHTTP(w, r)
-		case PluginHostServiceSetSettingProcedure:
-			pluginHostServiceSetSettingHandler.ServeHTTP(w, r)
-		default:
-			http.NotFound(w, r)
-		}
-	})
-}
-
-// UnimplementedPluginHostServiceHandler returns CodeUnimplemented from all methods.
-type UnimplementedPluginHostServiceHandler struct{}
-
-func (UnimplementedPluginHostServiceHandler) EmitEvent(context.Context, *connect.Request[v1.PluginHostServiceEmitEventRequest]) (*connect.Response[v1.PluginHostServiceEmitEventResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.EmitEvent is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) Log(context.Context, *connect.Request[v1.PluginHostServiceLogRequest]) (*connect.Response[v1.PluginHostServiceLogResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.Log is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) KVGet(context.Context, *connect.Request[v1.PluginHostServiceKVGetRequest]) (*connect.Response[v1.PluginHostServiceKVGetResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.KVGet is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) KVSet(context.Context, *connect.Request[v1.PluginHostServiceKVSetRequest]) (*connect.Response[v1.PluginHostServiceKVSetResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.KVSet is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) KVDelete(context.Context, *connect.Request[v1.PluginHostServiceKVDeleteRequest]) (*connect.Response[v1.PluginHostServiceKVDeleteResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.KVDelete is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) AddSessionStream(context.Context, *connect.Request[v1.PluginHostServiceAddSessionStreamRequest]) (*connect.Response[v1.PluginHostServiceAddSessionStreamResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.AddSessionStream is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) RemoveSessionStream(context.Context, *connect.Request[v1.PluginHostServiceRemoveSessionStreamRequest]) (*connect.Response[v1.PluginHostServiceRemoveSessionStreamResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.RemoveSessionStream is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) JoinFocus(context.Context, *connect.Request[v1.PluginHostServiceJoinFocusRequest]) (*connect.Response[v1.PluginHostServiceJoinFocusResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.JoinFocus is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) LeaveFocus(context.Context, *connect.Request[v1.PluginHostServiceLeaveFocusRequest]) (*connect.Response[v1.PluginHostServiceLeaveFocusResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.LeaveFocus is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) LeaveFocusByTarget(context.Context, *connect.Request[v1.PluginHostServiceLeaveFocusByTargetRequest]) (*connect.Response[v1.PluginHostServiceLeaveFocusByTargetResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.LeaveFocusByTarget is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) PresentFocus(context.Context, *connect.Request[v1.PluginHostServicePresentFocusRequest]) (*connect.Response[v1.PluginHostServicePresentFocusResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.PresentFocus is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) QueryStreamHistory(context.Context, *connect.Request[v1.PluginHostServiceQueryStreamHistoryRequest]) (*connect.Response[v1.PluginHostServiceQueryStreamHistoryResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.QueryStreamHistory is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) DecryptOwnAuditRows(context.Context, *connect.Request[v1.DecryptOwnAuditRowsRequest]) (*connect.Response[v1.DecryptOwnAuditRowsResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.DecryptOwnAuditRows is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) RequestEmitToken(context.Context, *connect.Request[v1.PluginHostServiceRequestEmitTokenRequest]) (*connect.Response[v1.PluginHostServiceRequestEmitTokenResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.RequestEmitToken is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) SetConnectionFocus(context.Context, *connect.Request[v1.PluginHostServiceSetConnectionFocusRequest]) (*connect.Response[v1.PluginHostServiceSetConnectionFocusResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.SetConnectionFocus is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) GetConnectionFocus(context.Context, *connect.Request[v1.PluginHostServiceGetConnectionFocusRequest]) (*connect.Response[v1.PluginHostServiceGetConnectionFocusResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.GetConnectionFocus is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) AutoFocusOnJoin(context.Context, *connect.Request[v1.PluginHostServiceAutoFocusOnJoinRequest]) (*connect.Response[v1.PluginHostServiceAutoFocusOnJoinResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.AutoFocusOnJoin is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) IsAnyConnFocused(context.Context, *connect.Request[v1.PluginHostServiceIsAnyConnFocusedRequest]) (*connect.Response[v1.PluginHostServiceIsAnyConnFocusedResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.IsAnyConnFocused is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) Evaluate(context.Context, *connect.Request[v1.PluginHostServiceEvaluateRequest]) (*connect.Response[v1.PluginHostServiceEvaluateResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.Evaluate is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) ListCommands(context.Context, *connect.Request[v1.PluginHostServiceListCommandsRequest]) (*connect.Response[v1.PluginHostServiceListCommandsResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.ListCommands is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) GetCommandHelp(context.Context, *connect.Request[v1.PluginHostServiceGetCommandHelpRequest]) (*connect.Response[v1.PluginHostServiceGetCommandHelpResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.GetCommandHelp is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) GetSetting(context.Context, *connect.Request[v1.PluginHostServiceGetSettingRequest]) (*connect.Response[v1.PluginHostServiceGetSettingResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.GetSetting is not implemented"))
-}
-
-func (UnimplementedPluginHostServiceHandler) SetSetting(context.Context, *connect.Request[v1.PluginHostServiceSetSettingRequest]) (*connect.Response[v1.PluginHostServiceSetSettingResponse], error) {
-	return nil, connect.NewError(connect.CodeUnimplemented, errors.New("holomush.plugin.v1.PluginHostService.SetSetting is not implemented"))
 }
