+++ test/integration/plugin/testdata/cmd_introspection_plugin/main.go
@@ -3,7 +3,7 @@
 
 // Package main implements the cmd_introspection_plugin: a test-only binary
 // plugin used by command_introspection_parity_test.go to prove INV-COMMAND-2
-// (runtime parity): the binary PluginHostService.ListCommands and the Lua
+// (runtime parity): the binary host.v1 CommandRegistryService.ListCommands and the Lua
 // holomush.list_commands host function delegate to the same
 // commandquery.Querier and return identical filtered command-name sets.
 //
