+++ web/src/lib/connect/holomush/plugin/v1/plugin_connect.ts
@@ -8,16 +8,16 @@
 /* eslint-disable */
 // @ts-nocheck
 
-import { HandleCommandRequest, HandleCommandResponse, HandleEventRequest, HandleEventResponse, InitRequest, InitResponse, PluginHostServiceAddSessionStreamRequest, PluginHostServiceAddSessionStreamResponse, PluginHostServiceAutoFocusOnJoinRequest, PluginHostServiceAutoFocusOnJoinResponse, PluginHostServiceEmitEventRequest, PluginHostServiceEmitEventResponse, PluginHostServiceEvaluateRequest, PluginHostServiceEvaluateResponse, PluginHostServiceGetCommandHelpRequest, PluginHostServiceGetCommandHelpResponse, PluginHostServiceGetConnectionFocusRequest, PluginHostServiceGetConnectionFocusResponse, PluginHostServiceGetSettingRequest, PluginHostServiceGetSettingResponse, PluginHostServiceIsAnyConnFocusedRequest, PluginHostServiceIsAnyConnFocusedResponse, PluginHostServiceJoinFocusRequest, PluginHostServiceJoinFocusResponse, PluginHostServiceKVDeleteRequest, PluginHostServiceKVDeleteResponse, PluginHostServiceKVGetRequest, PluginHostServiceKVGetResponse, PluginHostServiceKVSetRequest, PluginHostServiceKVSetResponse, PluginHostServiceLeaveFocusByTargetRequest, PluginHostServiceLeaveFocusByTargetResponse, PluginHostServiceLeaveFocusRequest, PluginHostServiceLeaveFocusResponse, PluginHostServiceListCommandsRequest, PluginHostServiceListCommandsResponse, PluginHostServiceLogRequest, PluginHostServiceLogResponse, PluginHostServicePresentFocusRequest, PluginHostServicePresentFocusResponse, PluginHostServiceQueryStreamHistoryRequest, PluginHostServiceQueryStreamHistoryResponse, PluginHostServiceRemoveSessionStreamRequest, PluginHostServiceRemoveSessionStreamResponse, PluginHostServiceRequestEmitTokenRequest, PluginHostServiceRequestEmitTokenResponse, PluginHostServiceSetConnectionFocusRequest, PluginHostServiceSetConnectionFocusResponse, PluginHostServiceSetSettingRequest, PluginHostServiceSetSettingResponse, QuerySessionStreamsRequest, QuerySessionStreamsResponse } from "./plugin_pb.js";
+import { HandleCommandRequest, HandleCommandResponse, HandleEventRequest, HandleEventResponse, InitRequest, InitResponse, QuerySessionStreamsRequest, QuerySessionStreamsResponse } from "./plugin_pb.js";
 import { MethodKind } from "@bufbuild/protobuf";
-import { DecryptOwnAuditRowsRequest, DecryptOwnAuditRowsResponse } from "./audit_pb.js";
 
 /**
  * PluginService is the plugin-IMPLEMENTED half of the binary-plugin contract:
  * the gRPC server runs inside the plugin subprocess (hashicorp/go-plugin) and
  * the host dials it. The host adapter is pkg/plugin.pluginServerAdapter, which
- * bridges these RPCs to a plugin author's Go handler. Contrast PluginHostService
- * below, which the host implements and plugins call back into.
+ * bridges these RPCs to a plugin author's Go handler. The host-implemented
+ * callback surface plugins dial in return is the per-capability
+ * holomush.plugin.host.v1 services (api/proto/holomush/plugin/host/v1).
  *
  * @generated from service holomush.plugin.v1.PluginService
  */
@@ -84,379 +84,6 @@ export const PluginService = {
       name: "QuerySessionStreams",
       I: QuerySessionStreamsRequest,
       O: QuerySessionStreamsResponse,
-      kind: MethodKind.Unary,
-    },
-  }
-} as const;
-
-/**
- * PluginHostService is the host-IMPLEMENTED half of the contract: the server
- * runs in the host process and binary plugins dial it to call back for event
- * emission, ABAC evaluation, audit decryption, focus mutation, and history
- * reads. Registered live at internal/plugin/goplugin/host_service.go
- * (RegisterPluginHostServiceServer, struct pluginHostServiceServer) — this
- * surface IS served in production.
- *
- * NOTE: the registered server embeds UnimplementedPluginHostServiceServer and
- * implements only 12 of the 18 RPCs below. Log, KVGet, KVSet, KVDelete,
- * AddSessionStream, and RemoveSessionStream are declared here but have no
- * server impl and no production client — they return codes.Unimplemented
- * (tracked in holomush-l6std). Their comments below document that unwired
- * reality, not aspirational behavior.
- *
- * @generated from service holomush.plugin.v1.PluginHostService
- */
-export const PluginHostService = {
-  typeName: "holomush.plugin.v1.PluginHostService",
-  methods: {
-    /**
-     * EmitEvent publishes one plugin-originated event onto the bus through the
-     * host emit fence. SERVED: pluginHostServiceServer.EmitEvent. The caller's
-     * identity is NOT trusted from the wire — the plugin presents a host-issued
-     * dispatch token in the x-holomush-emit-token metadata header, the host
-     * recovers the vouched-for actor from tokenStore.Lookup(pluginName, token),
-     * and a missing/foreign token is rejected (EMIT_TOKEN_MISSING /
-     * EMIT_TOKEN_REJECTED). The recovered actor then flows through
-     * PluginEventEmitter.Emit, which enforces the manifest gates: emits (subject
-     * namespace must be declared), actor_kinds_claimable (actor kind must be
-     * listed — EMIT_ACTOR_KIND_NOT_CLAIMABLE), and the crypto.emits sensitivity
-     * fence. These gates fire identically for Lua and binary plugins (plugin
-     * runtime symmetry); the token mechanism is the binary-side forgery fence.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.EmitEvent
-     */
-    emitEvent: {
-      name: "EmitEvent",
-      I: PluginHostServiceEmitEventRequest,
-      O: PluginHostServiceEmitEventResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * Log forwards a plugin log line to the host logger. DECLARED BUT UNSERVED:
-     * pluginHostServiceServer does not implement it and no production client
-     * calls it, so it currently returns codes.Unimplemented (holomush-l6std).
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.Log
-     */
-    log: {
-      name: "Log",
-      I: PluginHostServiceLogRequest,
-      O: PluginHostServiceLogResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * KVGet reads a value from the plugin's namespaced key-value store.
-     * DECLARED BUT UNSERVED (holomush-l6std): no server impl, no client; returns
-     * codes.Unimplemented today.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.KVGet
-     */
-    kVGet: {
-      name: "KVGet",
-      I: PluginHostServiceKVGetRequest,
-      O: PluginHostServiceKVGetResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * KVSet writes a value into the plugin's namespaced key-value store.
-     * DECLARED BUT UNSERVED (holomush-l6std): no server impl, no client; returns
-     * codes.Unimplemented today.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.KVSet
-     */
-    kVSet: {
-      name: "KVSet",
-      I: PluginHostServiceKVSetRequest,
-      O: PluginHostServiceKVSetResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * KVDelete removes a key from the plugin's namespaced key-value store.
-     * DECLARED BUT UNSERVED (holomush-l6std): no server impl, no client; returns
-     * codes.Unimplemented today.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.KVDelete
-     */
-    kVDelete: {
-      name: "KVDelete",
-      I: PluginHostServiceKVDeleteRequest,
-      O: PluginHostServiceKVDeleteResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * AddSessionStream subscribes an active session to an additional stream
-     * mid-session. DECLARED BUT UNSERVED (holomush-l6std): no server impl, no
-     * production client; returns codes.Unimplemented today. (The wire comment
-     * about SESSION_NOT_FOUND describes the intended-but-unimplemented contract.)
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.AddSessionStream
-     */
-    addSessionStream: {
-      name: "AddSessionStream",
-      I: PluginHostServiceAddSessionStreamRequest,
-      O: PluginHostServiceAddSessionStreamResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * RemoveSessionStream unsubscribes an active session from a stream.
-     * DECLARED BUT UNSERVED (holomush-l6std): no server impl, no production
-     * client; returns codes.Unimplemented today.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.RemoveSessionStream
-     */
-    removeSessionStream: {
-      name: "RemoveSessionStream",
-      I: PluginHostServiceRemoveSessionStreamRequest,
-      O: PluginHostServiceRemoveSessionStreamResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * JoinFocus adds a focus membership (e.g. a scene) to a session via the
-     * host focus coordinator. SERVED: pluginHostServiceServer.JoinFocus. The
-     * plugin declares intent; the coordinator applies the kind-specific replay
-     * policy. Fails if the focus coordinator is not configured.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.JoinFocus
-     */
-    joinFocus: {
-      name: "JoinFocus",
-      I: PluginHostServiceJoinFocusRequest,
-      O: PluginHostServiceJoinFocusResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * LeaveFocus removes one focus membership from a session. SERVED:
-     * pluginHostServiceServer.LeaveFocus. Idempotent — leaving a target the
-     * session does not hold is a successful no-op.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.LeaveFocus
-     */
-    leaveFocus: {
-      name: "LeaveFocus",
-      I: PluginHostServiceLeaveFocusRequest,
-      O: PluginHostServiceLeaveFocusResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * LeaveFocusByTarget removes the given focus membership from every
-     * non-expired session that holds it — cross-session fan-out (e.g. a
-     * scene-end reaching all participants). SERVED:
-     * pluginHostServiceServer.LeaveFocusByTarget. Partial success is normal:
-     * per-session failures are aggregated into the response (succeeded /
-     * total_scanned / failed_session_ids) rather than aborting the sweep; the
-     * RPC error is reserved for the enumeration step itself failing.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.LeaveFocusByTarget
-     */
-    leaveFocusByTarget: {
-      name: "LeaveFocusByTarget",
-      I: PluginHostServiceLeaveFocusByTargetRequest,
-      O: PluginHostServiceLeaveFocusByTargetResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * PresentFocus repoints a session's PresentingFocus to an existing
-     * membership. SERVED: pluginHostServiceServer.PresentFocus. The target MUST
-     * already be in the session's FocusMemberships (membership is validated, not
-     * implicitly created).
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.PresentFocus
-     */
-    presentFocus: {
-      name: "PresentFocus",
-      I: PluginHostServicePresentFocusRequest,
-      O: PluginHostServicePresentFocusResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * QueryStreamHistory reads the tail of a stream for plugin-side display,
-     * backward-paginated by opaque cursor. SERVED:
-     * pluginHostServiceServer.QueryStreamHistory via HistoryReader.ReplayTail.
-     * Read-only: it does not advance session cursors or mutate session state.
-     * A negative count is rejected (INVALID_ARGUMENT); count is CLAMPED to
-     * maxQueryStreamHistoryCount (500), not rejected, when too large.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.QueryStreamHistory
-     */
-    queryStreamHistory: {
-      name: "QueryStreamHistory",
-      I: PluginHostServiceQueryStreamHistoryRequest,
-      O: PluginHostServiceQueryStreamHistoryResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * DecryptOwnAuditRows decrypts a batch of the calling plugin's OWN encrypted
-     * audit rows host-side; the plugin never holds a DEK. SERVED:
-     * pluginHostServiceServer.DecryptOwnAuditRows via ReadbackDecryptor.
-     * DecryptOwnRows. Authorization is two-gate (INV-CRYPTO-27): OwnerMap subject
-     * ownership (g1) plus the crypto.emits[].readback manifest flag (g2). Each
-     * input row gets an independent RowResult (INV-CRYPTO-37) carrying either
-     * plaintext or a stable snake_case no_plaintext_reason ("not_owner",
-     * "auth_guard_deny", "dek_missing", "downgrade_refused", "stale_dek",
-     * "audit_queue_full", "internal"; readback.go reasonToWire). Request /
-     * response shapes and RowResult live in audit.proto (AuditRow domain).
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.DecryptOwnAuditRows
-     */
-    decryptOwnAuditRows: {
-      name: "DecryptOwnAuditRows",
-      I: DecryptOwnAuditRowsRequest,
-      O: DecryptOwnAuditRowsResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * RequestEmitToken issues a self-token bound to {ActorPlugin, pluginName} so
-     * a plugin-served gRPC handler (e.g. SceneService.CreateScene) — which is NOT
-     * reached via DeliverEvent/DeliverCommand and so holds no dispatch token —
-     * can still call EmitEvent. SERVED: pluginHostServiceServer.RequestEmitToken.
-     * The plugin's identity is taken from the mTLS-bound server struct
-     * (s.pluginName); the request carries no identity fields, so a plugin cannot
-     * impersonate another actor or escalate to a character actor through this
-     * RPC. The actor_kinds_claimable manifest gate still fires when the issued
-     * token is later spent at EmitEvent. (Spec §3.3.5 / §5.4 two-token pattern.)
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.RequestEmitToken
-     */
-    requestEmitToken: {
-      name: "RequestEmitToken",
-      I: PluginHostServiceRequestEmitTokenRequest,
-      O: PluginHostServiceRequestEmitTokenResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * SetConnectionFocus is the Phase-5 explicit focus mutation for a single
-     * Connection. SERVED: pluginHostServiceServer.SetConnectionFocus. The
-     * substrate validates the requested membership against the session's
-     * FocusMemberships (D4), then writes Connection.FocusKey and (D9-gated)
-     * Info.PresentingFocus atomically under one Store-lock acquisition (D7).
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.SetConnectionFocus
-     */
-    setConnectionFocus: {
-      name: "SetConnectionFocus",
-      I: PluginHostServiceSetConnectionFocusRequest,
-      O: PluginHostServiceSetConnectionFocusResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * GetConnectionFocus returns the named connection's current per-connection
-     * focus, or absent when the connection is grid-focused (FocusKey nil) or
-     * unknown. Read-only counterpart of SetConnectionFocus; lets plugins route
-     * connection-scoped operations (e.g. scene pose) to the focused target.
-     * See goplugin/host_service.go::GetConnectionFocus.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.GetConnectionFocus
-     */
-    getConnectionFocus: {
-      name: "GetConnectionFocus",
-      I: PluginHostServiceGetConnectionFocusRequest,
-      O: PluginHostServiceGetConnectionFocusResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * AutoFocusOnJoin is the Phase-5 fan-out that focuses all of a character's
-     * terminal/telnet connections on a scene at once. SERVED:
-     * pluginHostServiceServer.AutoFocusOnJoin. Connections already explicitly
-     * focused elsewhere are skipped (D8). The caller MUST have completed
-     * JoinFocus first, since the substrate requires the membership to exist.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.AutoFocusOnJoin
-     */
-    autoFocusOnJoin: {
-      name: "AutoFocusOnJoin",
-      I: PluginHostServiceAutoFocusOnJoinRequest,
-      O: PluginHostServiceAutoFocusOnJoinResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * IsAnyConnFocused is the Phase-5 notification-emission helper: it reports
-     * whether any of the character's connections currently focuses the given
-     * scene, so callers can decide whether to emit a focus-related notification.
-     * SERVED: pluginHostServiceServer.IsAnyConnFocused.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.IsAnyConnFocused
-     */
-    isAnyConnFocused: {
-      name: "IsAnyConnFocused",
-      I: PluginHostServiceIsAnyConnFocusedRequest,
-      O: PluginHostServiceIsAnyConnFocusedResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * Evaluate runs the host ABAC engine for one action against one resource
-     * instance owned by the calling plugin. SERVED:
-     * pluginHostServiceServer.Evaluate. The subject is derived host-side from the
-     * dispatch token exactly as EmitEvent does (token→actor recovery) — there is
-     * no subject field on the wire (spec §2, INV-PLUGIN-22). Fails closed on nil engine,
-     * missing/rejected token, empty actor subject, or a resource type the plugin
-     * does not own.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.Evaluate
-     */
-    evaluate: {
-      name: "Evaluate",
-      I: PluginHostServiceEvaluateRequest,
-      O: PluginHostServiceEvaluateResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * ListCommands enumerates the commands the named character may execute,
-     * ABAC-filtered by the host. SERVED: pluginHostServiceServer.ListCommands,
-     * delegating to commandquery.Querier.Available. The subject is the request's
-     * character_id (parity with the Lua holomush.list_commands(character_id) host
-     * function — not the dispatch-token actor, since this is read-only metadata,
-     * not an actor-gated mutation). incomplete is true when engine errors hid
-     * some commands.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.ListCommands
-     */
-    listCommands: {
-      name: "ListCommands",
-      I: PluginHostServiceListCommandsRequest,
-      O: PluginHostServiceListCommandsResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * GetCommandHelp returns full help detail for one command after an access
-     * check for character_id. SERVED: pluginHostServiceServer.GetCommandHelp,
-     * delegating to commandquery.Querier.Help. Mirrors the Lua
-     * holomush.get_command_help(name, character_id) host function.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.GetCommandHelp
-     */
-    getCommandHelp: {
-      name: "GetCommandHelp",
-      I: PluginHostServiceGetCommandHelpRequest,
-      O: PluginHostServiceGetCommandHelpResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * GetSetting reads a single-scope setting in the calling plugin's owner
-     * partition (owner bound host-side from the authenticated plugin name, never
-     * from the request). The handler resolves SettingScope to its backing store;
-     * a missing key returns a successful response with found=false, never a
-     * codes.NotFound status error.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.GetSetting
-     */
-    getSetting: {
-      name: "GetSetting",
-      I: PluginHostServiceGetSettingRequest,
-      O: PluginHostServiceGetSettingResponse,
-      kind: MethodKind.Unary,
-    },
-    /**
-     * SetSetting writes a single-scope setting in the calling plugin's partition;
-     * GAME scope requires an operator authorization decision (host-enforced, not
-     * trusted from the wire). The owner partition is bound host-side from the
-     * authenticated plugin name.
-     *
-     * @generated from rpc holomush.plugin.v1.PluginHostService.SetSetting
-     */
-    setSetting: {
-      name: "SetSetting",
-      I: PluginHostServiceSetSettingRequest,
-      O: PluginHostServiceSetSettingResponse,
       kind: MethodKind.Unary,
     },
   }
