+++ web/src/lib/connect/holomush/plugin/v1/plugin_pb.ts
@@ -10,15 +10,13 @@
 import type { GenEnum, GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv2";
 import { enumDesc, fileDesc, messageDesc, serviceDesc } from "@bufbuild/protobuf/codegenv2";
 import { file_buf_validate_validate } from "../../../buf/validate/validate_pb";
-import type { DecryptOwnAuditRowsRequestSchema, DecryptOwnAuditRowsResponseSchema } from "./audit_pb";
-import { file_holomush_plugin_v1_audit } from "./audit_pb";
 import type { Message } from "@bufbuild/protobuf";
 
 /**
  * Describes the file holomush/plugin/v1/plugin.proto.
  */
 export const file_holomush_plugin_v1_plugin: GenFile = /*@__PURE__*/
-  fileDesc("Ch9ob2xvbXVzaC9wbHVnaW4vdjEvcGx1Z2luLnByb3RvEhJob2xvbXVzaC5wbHVnaW4udjEiuAIKDVNlcnZpY2VDb25maWcSGQoRY29ubmVjdGlvbl9zdHJpbmcYASABKAkSUgoRcmVxdWlyZWRfc2VydmljZXMYAiADKAsyNy5ob2xvbXVzaC5wbHVnaW4udjEuU2VydmljZUNvbmZpZy5SZXF1aXJlZFNlcnZpY2VzRW50cnkSSgoNcGx1Z2luX2NvbmZpZxgDIAMoCzIzLmhvbG9tdXNoLnBsdWdpbi52MS5TZXJ2aWNlQ29uZmlnLlBsdWdpbkNvbmZpZ0VudHJ5GjcKFVJlcXVpcmVkU2VydmljZXNFbnRyeRILCgNrZXkYASABKAkSDQoFdmFsdWUYAiABKAk6AjgBGjMKEVBsdWdpbkNvbmZpZ0VudHJ5EgsKA2tleRgBIAEoCRINCgV2YWx1ZRgCIAEoCToCOAEiQAoLSW5pdFJlcXVlc3QSMQoGY29uZmlnGAEgASgLMiEuaG9sb211c2gucGx1Z2luLnYxLlNlcnZpY2VDb25maWciSAoMSW5pdFJlc3BvbnNlEhkKEXByb3ZpZGVkX3NlcnZpY2VzGAEgAygJEh0KFXJlZ2lzdGVyZWRfZW1pdF90eXBlcxgCIAMoCSLDAQoFRXZlbnQSEwoCaWQYASABKAlCB7pIBHICEAESFwoGc3RyZWFtGAIgASgJQge6SARyAhABEhUKBHR5cGUYAyABKAlCB7pIBHICEAESEQoJdGltZXN0YW1wGAQgASgDEhsKCmFjdG9yX2tpbmQYBSABKAlCB7pIBHICEAESGQoIYWN0b3JfaWQYBiABKAlCB7pIBHICEAESGgoHcGF5bG9hZBgHIAEoCUIJukgGcgQYgIAEEg4KBmN1cnNvchgIIAEoDCJqCglFbWl0RXZlbnQSFwoGc3RyZWFtGAEgASgJQge6SARyAhABEhUKBHR5cGUYAiABKAlCB7pIBHICEAESGgoHcGF5bG9hZBgDIAEoCUIJukgGcgQYgIAEEhEKCXNlbnNpdGl2ZRgEIAEoCCI+ChJIYW5kbGVFdmVudFJlcXVlc3QSKAoFZXZlbnQYASABKAsyGS5ob2xvbXVzaC5wbHVnaW4udjEuRXZlbnQiSQoTSGFuZGxlRXZlbnRSZXNwb25zZRIyCgtlbWl0X2V2ZW50cxgBIAMoCzIdLmhvbG9tdXNoLnBsdWdpbi52MS5FbWl0RXZlbnQijQIKDkNvbW1hbmRSZXF1ZXN0EhgKB2NvbW1hbmQYASABKAlCB7pIBHICEAESFgoEYXJncxgCIAEoCUIIukgFcgMYgEASGwoJcmF3X2lucHV0GAMgASgJQgi6SAVyAxiAQBIdCgxjaGFyYWN0ZXJfaWQYBCABKAlCB7pIBHICEAESHwoOY2hhcmFjdGVyX25hbWUYBSABKAlCB7pIBHICEAESHAoLbG9jYXRpb25faWQYBiABKAlCB7pIBHICEAESGwoKc2Vzc2lvbl9pZBgHIAEoCUIHukgEcgIQARIaCglwbGF5ZXJfaWQYCCABKAlCB7pIBHICEAESFQoNY29ubmVjdGlvbl9pZBgJIAEoCSLJAQoPQ29tbWFuZFJlc3BvbnNlEjEKBnN0YXR1cxgBIAEoDjIhLmhvbG9tdXNoLnBsdWdpbi52MS5Db21tYW5kU3RhdHVzEhgKBm91dHB1dBgCIAEoCUIIukgFcgMYgEASLQoGZXZlbnRzGAMgAygLMh0uaG9sb211c2gucGx1Z2luLnYxLkVtaXRFdmVudBI6CgthdWRpdF9oaW50cxgEIAMoCzIlLmhvbG9tdXNoLnBsdWdpbi52MS5BdWRpdERlY2lzaW9uSGludCLMAgoRQXVkaXREZWNpc2lvbkhpbnQSFgoCaWQYASABKAlCCrpIB3IFEAEYgAESFgoEbmFtZRgCIAEoCUIIukgFcgMYgAISGQoHbWVzc2FnZRgDIAEoCUIIukgFcgMYgAgSLwoGZWZmZWN0GAQgASgOMh8uaG9sb211c2gucGx1Z2luLnYxLkF1ZGl0RWZmZWN0EiEKEGFjdGlvbl9xdWFsaWZpZXIYBSABKAlCB7pIBHICGEASGgoIcmVzb3VyY2UYBiABKAlCCLpIBXIDGIACEkkKCmF0dHJpYnV0ZXMYByADKAsyNS5ob2xvbXVzaC5wbHVnaW4udjEuQXVkaXREZWNpc2lvbkhpbnQuQXR0cmlidXRlc0VudHJ5GjEKD0F0dHJpYnV0ZXNFbnRyeRILCgNrZXkYASABKAkSDQoFdmFsdWUYAiABKAk6AjgBIksKFEhhbmRsZUNvbW1hbmRSZXF1ZXN0EjMKB2NvbW1hbmQYASABKAsyIi5ob2xvbXVzaC5wbHVnaW4udjEuQ29tbWFuZFJlcXVlc3QiTgoVSGFuZGxlQ29tbWFuZFJlc3BvbnNlEjUKCHJlc3BvbnNlGAEgASgLMiMuaG9sb211c2gucGx1Z2luLnYxLkNvbW1hbmRSZXNwb25zZSJ9CiFQbHVnaW5Ib3N0U2VydmljZUVtaXRFdmVudFJlcXVlc3QSFwoGc3RyZWFtGAEgASgJQge6SARyAhABEhsKCmV2ZW50X3R5cGUYAiABKAlCB7pIBHICEAESDwoHcGF5bG9hZBgDIAEoDBIRCglzZW5zaXRpdmUYBCABKAgiJAoiUGx1Z2luSG9zdFNlcnZpY2VFbWl0RXZlbnRSZXNwb25zZSJQChtQbHVnaW5Ib3N0U2VydmljZUxvZ1JlcXVlc3QSFgoFbGV2ZWwYASABKAlCB7pIBHICEAESGQoHbWVzc2FnZRgCIAEoCUIIukgFcgMYgEAiHgocUGx1Z2luSG9zdFNlcnZpY2VMb2dSZXNwb25zZSJTCh1QbHVnaW5Ib3N0U2VydmljZUtWR2V0UmVxdWVzdBIcCgtwbHVnaW5fbmFtZRgBIAEoCUIHukgEcgIQARIUCgNrZXkYAiABKAlCB7pIBHICEAEiPgoeUGx1Z2luSG9zdFNlcnZpY2VLVkdldFJlc3BvbnNlEg0KBXZhbHVlGAEgASgJEg0KBWZvdW5kGAIgASgIImIKHVBsdWdpbkhvc3RTZXJ2aWNlS1ZTZXRSZXF1ZXN0EhwKC3BsdWdpbl9uYW1lGAEgASgJQge6SARyAhABEhQKA2tleRgCIAEoCUIHukgEcgIQARINCgV2YWx1ZRgDIAEoCSIgCh5QbHVnaW5Ib3N0U2VydmljZUtWU2V0UmVzcG9uc2UiVgogUGx1Z2luSG9zdFNlcnZpY2VLVkRlbGV0ZVJlcXVlc3QSHAoLcGx1Z2luX25hbWUYASABKAlCB7pIBHICEAESFAoDa2V5GAIgASgJQge6SARyAhABIiMKIVBsdWdpbkhvc3RTZXJ2aWNlS1ZEZWxldGVSZXNwb25zZSKbAQooUGx1Z2luSG9zdFNlcnZpY2VBZGRTZXNzaW9uU3RyZWFtUmVxdWVzdBIbCgpzZXNzaW9uX2lkGAEgASgJQge6SARyAhABEhcKBnN0cmVhbRgCIAEoCUIHukgEcgIQARI5CgtyZXBsYXlfbW9kZRgDIAEoDjIkLmhvbG9tdXNoLnBsdWdpbi52MS5TdHJlYW1SZXBsYXlNb2RlIisKKVBsdWdpbkhvc3RTZXJ2aWNlQWRkU2Vzc2lvblN0cmVhbVJlc3BvbnNlImMKK1BsdWdpbkhvc3RTZXJ2aWNlUmVtb3ZlU2Vzc2lvblN0cmVhbVJlcXVlc3QSGwoKc2Vzc2lvbl9pZBgBIAEoCUIHukgEcgIQARIXCgZzdHJlYW0YAiABKAlCB7pIBHICEAEiLgosUGx1Z2luSG9zdFNlcnZpY2VSZW1vdmVTZXNzaW9uU3RyZWFtUmVzcG9uc2UiWQoaUXVlcnlTZXNzaW9uU3RyZWFtc1JlcXVlc3QSFAoMY2hhcmFjdGVyX2lkGAEgASgJEhEKCXBsYXllcl9pZBgCIAEoCRISCgpzZXNzaW9uX2lkGAMgASgJIj0KG1F1ZXJ5U2Vzc2lvblN0cmVhbXNSZXNwb25zZRIPCgdzdHJlYW1zGAEgAygJEg0KBWVycm9yGAIgASgJIlMKCEZvY3VzS2V5EisKBGtpbmQYASABKA4yHS5ob2xvbXVzaC5wbHVnaW4udjEuRm9jdXNLaW5kEhoKCXRhcmdldF9pZBgCIAEoCUIHukgEcgIQASJuCiFQbHVnaW5Ib3N0U2VydmljZUpvaW5Gb2N1c1JlcXVlc3QSGwoKc2Vzc2lvbl9pZBgBIAEoCUIHukgEcgIQARIsCgZ0YXJnZXQYAiABKAsyHC5ob2xvbXVzaC5wbHVnaW4udjEuRm9jdXNLZXkiJAoiUGx1Z2luSG9zdFNlcnZpY2VKb2luRm9jdXNSZXNwb25zZSJvCiJQbHVnaW5Ib3N0U2VydmljZUxlYXZlRm9jdXNSZXF1ZXN0EhsKCnNlc3Npb25faWQYASABKAlCB7pIBHICEAESLAoGdGFyZ2V0GAIgASgLMhwuaG9sb211c2gucGx1Z2luLnYxLkZvY3VzS2V5IiUKI1BsdWdpbkhvc3RTZXJ2aWNlTGVhdmVGb2N1c1Jlc3BvbnNlIloKKlBsdWdpbkhvc3RTZXJ2aWNlTGVhdmVGb2N1c0J5VGFyZ2V0UmVxdWVzdBIsCgZ0YXJnZXQYASABKAsyHC5ob2xvbXVzaC5wbHVnaW4udjEuRm9jdXNLZXkicworUGx1Z2luSG9zdFNlcnZpY2VMZWF2ZUZvY3VzQnlUYXJnZXRSZXNwb25zZRIRCglzdWNjZWVkZWQYASABKAUSFQoNdG90YWxfc2Nhbm5lZBgCIAEoBRIaChJmYWlsZWRfc2Vzc2lvbl9pZHMYAyADKAkicQokUGx1Z2luSG9zdFNlcnZpY2VQcmVzZW50Rm9jdXNSZXF1ZXN0EhsKCnNlc3Npb25faWQYASABKAlCB7pIBHICEAESLAoGdGFyZ2V0GAIgASgLMhwuaG9sb211c2gucGx1Z2luLnYxLkZvY3VzS2V5IicKJVBsdWdpbkhvc3RTZXJ2aWNlUHJlc2VudEZvY3VzUmVzcG9uc2UiewoqUGx1Z2luSG9zdFNlcnZpY2VRdWVyeVN0cmVhbUhpc3RvcnlSZXF1ZXN0EhcKBnN0cmVhbRgBIAEoCUIHukgEcgIQARINCgVjb3VudBgCIAEoBRIVCg1ub3RfYmVmb3JlX21zGAMgASgDEg4KBmN1cnNvchgEIAEoDCJtCitQbHVnaW5Ib3N0U2VydmljZVF1ZXJ5U3RyZWFtSGlzdG9yeVJlc3BvbnNlEikKBmV2ZW50cxgBIAMoCzIZLmhvbG9tdXNoLnBsdWdpbi52MS5FdmVudBITCgtuZXh0X2N1cnNvchgCIAEoDCIqCihQbHVnaW5Ib3N0U2VydmljZVJlcXVlc3RFbWl0VG9rZW5SZXF1ZXN0IjoKKVBsdWdpbkhvc3RTZXJ2aWNlUmVxdWVzdEVtaXRUb2tlblJlc3BvbnNlEg0KBXRva2VuGAEgASgJIp4BCipQbHVnaW5Ib3N0U2VydmljZVNldENvbm5lY3Rpb25Gb2N1c1JlcXVlc3QSFQoNY29ubmVjdGlvbl9pZBgBIAEoDBI0Cglmb2N1c19rZXkYAiABKAsyHC5ob2xvbXVzaC5wbHVnaW4udjEuRm9jdXNLZXlIAIgBARIVCg1pc19zY2VuZV9ncmlkGAMgASgIQgwKCl9mb2N1c19rZXkicQorUGx1Z2luSG9zdFNlcnZpY2VTZXRDb25uZWN0aW9uRm9jdXNSZXNwb25zZRI0Cglmb2N1c19rZXkYASABKAsyHC5ob2xvbXVzaC5wbHVnaW4udjEuRm9jdXNLZXlIAIgBAUIMCgpfZm9jdXNfa2V5IkMKKlBsdWdpbkhvc3RTZXJ2aWNlR2V0Q29ubmVjdGlvbkZvY3VzUmVxdWVzdBIVCg1jb25uZWN0aW9uX2lkGAEgASgMInEKK1BsdWdpbkhvc3RTZXJ2aWNlR2V0Q29ubmVjdGlvbkZvY3VzUmVzcG9uc2USNAoJZm9jdXNfa2V5GAEgASgLMhwuaG9sb211c2gucGx1Z2luLnYxLkZvY3VzS2V5SACIAQFCDAoKX2ZvY3VzX2tleSJRCidQbHVnaW5Ib3N0U2VydmljZUF1dG9Gb2N1c09uSm9pblJlcXVlc3QSFAoMY2hhcmFjdGVyX2lkGAEgASgMEhAKCHNjZW5lX2lkGAIgASgMIssBCihQbHVnaW5Ib3N0U2VydmljZUF1dG9Gb2N1c09uSm9pblJlc3BvbnNlEh4KFmZvY3VzZWRfY29ubmVjdGlvbl9pZHMYASADKAwSHgoWdG90YWxfY29ubmVjdGlvbl9jb3VudBgCIAEoDRIeChZza2lwcGVkX2Nvbm5lY3Rpb25faWRzGAMgAygMEj8KFWZhaWxlZF9jb25uZWN0aW9uX2lkcxgEIAMoCzIgLmhvbG9tdXNoLnBsdWdpbi52MS5Gb2N1c0ZhaWx1cmUiXQoMRm9jdXNGYWlsdXJlEhUKDWNvbm5lY3Rpb25faWQYASABKAwSNgoGcmVhc29uGAIgASgOMiYuaG9sb211c2gucGx1Z2luLnYxLkZvY3VzRmFpbHVyZVJlYXNvbiJSCihQbHVnaW5Ib3N0U2VydmljZUlzQW55Q29ubkZvY3VzZWRSZXF1ZXN0EhQKDGNoYXJhY3Rlcl9pZBgBIAEoDBIQCghzY2VuZV9pZBgCIAEoDCI8CilQbHVnaW5Ib3N0U2VydmljZUlzQW55Q29ubkZvY3VzZWRSZXNwb25zZRIPCgdmb2N1c2VkGAEgASgIIlYKIFBsdWdpbkhvc3RTZXJ2aWNlRXZhbHVhdGVSZXF1ZXN0EhcKBmFjdGlvbhgBIAEoCUIHukgEcgIQARIZCghyZXNvdXJjZRgCIAEoCUIHukgEcgIQAyJcCiFQbHVnaW5Ib3N0U2VydmljZUV2YWx1YXRlUmVzcG9uc2USDwoHYWxsb3dlZBgBIAEoCBIOCgZyZWFzb24YAiABKAkSFgoObWF0Y2hlZF9wb2xpY3kYAyABKAkiWQocUGx1Z2luSG9zdFNlcnZpY2VDb21tYW5kSW5mbxIMCgRuYW1lGAEgASgJEgwKBGhlbHAYAiABKAkSDQoFdXNhZ2UYAyABKAkSDgoGc291cmNlGAQgASgJIkUKJFBsdWdpbkhvc3RTZXJ2aWNlTGlzdENvbW1hbmRzUmVxdWVzdBIdCgxjaGFyYWN0ZXJfaWQYASABKAlCB7pIBHICEBoifwolUGx1Z2luSG9zdFNlcnZpY2VMaXN0Q29tbWFuZHNSZXNwb25zZRJCCghjb21tYW5kcxgBIAMoCzIwLmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZUNvbW1hbmRJbmZvEhIKCmluY29tcGxldGUYAiABKAgiXgomUGx1Z2luSG9zdFNlcnZpY2VHZXRDb21tYW5kSGVscFJlcXVlc3QSFQoEbmFtZRgBIAEoCUIHukgEcgIQARIdCgxjaGFyYWN0ZXJfaWQYAiABKAlCB7pIBHICEBoidwonUGx1Z2luSG9zdFNlcnZpY2VHZXRDb21tYW5kSGVscFJlc3BvbnNlEgwKBG5hbWUYASABKAkSDAoEaGVscBgCIAEoCRINCgV1c2FnZRgDIAEoCRIRCgloZWxwX3RleHQYBCABKAkSDgoGc291cmNlGAUgASgJIoEBCiJQbHVnaW5Ib3N0U2VydmljZUdldFNldHRpbmdSZXF1ZXN0Ei8KBXNjb3BlGAEgASgOMiAuaG9sb211c2gucGx1Z2luLnYxLlNldHRpbmdTY29wZRIUCgxwcmluY2lwYWxfaWQYAiABKAkSFAoDa2V5GAMgASgJQge6SARyAhABIl8KI1BsdWdpbkhvc3RTZXJ2aWNlR2V0U2V0dGluZ1Jlc3BvbnNlEg0KBWZvdW5kGAEgASgIEhMKC3N0cmluZ19saXN0GAIgAygJEhQKDHN0cmluZ192YWx1ZRgDIAEoCSKWAQoiUGx1Z2luSG9zdFNlcnZpY2VTZXRTZXR0aW5nUmVxdWVzdBIvCgVzY29wZRgBIAEoDjIgLmhvbG9tdXNoLnBsdWdpbi52MS5TZXR0aW5nU2NvcGUSFAoMcHJpbmNpcGFsX2lkGAIgASgJEhQKA2tleRgDIAEoCUIHukgEcgIQARITCgtzdHJpbmdfbGlzdBgEIAMoCSIlCiNQbHVnaW5Ib3N0U2VydmljZVNldFNldHRpbmdSZXNwb25zZSqWAQoNQ29tbWFuZFN0YXR1cxIeChpDT01NQU5EX1NUQVRVU19VTlNQRUNJRklFRBAAEhUKEUNPTU1BTkRfU1RBVFVTX09LEAESGAoUQ09NTUFORF9TVEFUVVNfRVJST1IQAhIaChZDT01NQU5EX1NUQVRVU19GQUlMVVJFEAMSGAoUQ09NTUFORF9TVEFUVVNfRkFUQUwQBCpaCgtBdWRpdEVmZmVjdBIcChhBVURJVF9FRkZFQ1RfVU5TUEVDSUZJRUQQABIVChFBVURJVF9FRkZFQ1RfREVOWRABEhYKEkFVRElUX0VGRkVDVF9BTExPVxACKj0KCUZvY3VzS2luZBIaChZGT0NVU19LSU5EX1VOU1BFQ0lGSUVEEAASFAoQRk9DVVNfS0lORF9TQ0VORRABKnwKEFN0cmVhbVJlcGxheU1vZGUSIgoeU1RSRUFNX1JFUExBWV9NT0RFX1VOU1BFQ0lGSUVEEAASIgoeU1RSRUFNX1JFUExBWV9NT0RFX0ZST01fQ1VSU09SEAESIAocU1RSRUFNX1JFUExBWV9NT0RFX0xJVkVfT05MWRACKpUBChJGb2N1c0ZhaWx1cmVSZWFzb24SJAogRk9DVVNfRkFJTFVSRV9SRUFTT05fVU5TUEVDSUZJRUQQABIqCiZGT0NVU19GQUlMVVJFX1JFQVNPTl9NRU1CRVJTSElQX0FCU0VOVBABEi0KKUZPQ1VTX0ZBSUxVUkVfUkVBU09OX0NPTk5FQ1RJT05fTk9UX0ZPVU5EEAIqfAoMU2V0dGluZ1Njb3BlEh0KGVNFVFRJTkdfU0NPUEVfVU5TUEVDSUZJRUQQABIWChJTRVRUSU5HX1NDT1BFX0dBTUUQARIYChRTRVRUSU5HX1NDT1BFX1BMQVlFUhACEhsKF1NFVFRJTkdfU0NPUEVfQ0hBUkFDVEVSEAMymAMKDVBsdWdpblNlcnZpY2USSQoESW5pdBIfLmhvbG9tdXNoLnBsdWdpbi52MS5Jbml0UmVxdWVzdBogLmhvbG9tdXNoLnBsdWdpbi52MS5Jbml0UmVzcG9uc2USXgoLSGFuZGxlRXZlbnQSJi5ob2xvbXVzaC5wbHVnaW4udjEuSGFuZGxlRXZlbnRSZXF1ZXN0GicuaG9sb211c2gucGx1Z2luLnYxLkhhbmRsZUV2ZW50UmVzcG9uc2USZAoNSGFuZGxlQ29tbWFuZBIoLmhvbG9tdXNoLnBsdWdpbi52MS5IYW5kbGVDb21tYW5kUmVxdWVzdBopLmhvbG9tdXNoLnBsdWdpbi52MS5IYW5kbGVDb21tYW5kUmVzcG9uc2USdgoTUXVlcnlTZXNzaW9uU3RyZWFtcxIuLmhvbG9tdXNoLnBsdWdpbi52MS5RdWVyeVNlc3Npb25TdHJlYW1zUmVxdWVzdBovLmhvbG9tdXNoLnBsdWdpbi52MS5RdWVyeVNlc3Npb25TdHJlYW1zUmVzcG9uc2UylBgKEVBsdWdpbkhvc3RTZXJ2aWNlEnoKCUVtaXRFdmVudBI1LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZUVtaXRFdmVudFJlcXVlc3QaNi5ob2xvbXVzaC5wbHVnaW4udjEuUGx1Z2luSG9zdFNlcnZpY2VFbWl0RXZlbnRSZXNwb25zZRJoCgNMb2cSLy5ob2xvbXVzaC5wbHVnaW4udjEuUGx1Z2luSG9zdFNlcnZpY2VMb2dSZXF1ZXN0GjAuaG9sb211c2gucGx1Z2luLnYxLlBsdWdpbkhvc3RTZXJ2aWNlTG9nUmVzcG9uc2USbgoFS1ZHZXQSMS5ob2xvbXVzaC5wbHVnaW4udjEuUGx1Z2luSG9zdFNlcnZpY2VLVkdldFJlcXVlc3QaMi5ob2xvbXVzaC5wbHVnaW4udjEuUGx1Z2luSG9zdFNlcnZpY2VLVkdldFJlc3BvbnNlEm4KBUtWU2V0EjEuaG9sb211c2gucGx1Z2luLnYxLlBsdWdpbkhvc3RTZXJ2aWNlS1ZTZXRSZXF1ZXN0GjIuaG9sb211c2gucGx1Z2luLnYxLlBsdWdpbkhvc3RTZXJ2aWNlS1ZTZXRSZXNwb25zZRJ3CghLVkRlbGV0ZRI0LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZUtWRGVsZXRlUmVxdWVzdBo1LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZUtWRGVsZXRlUmVzcG9uc2USjwEKEEFkZFNlc3Npb25TdHJlYW0SPC5ob2xvbXVzaC5wbHVnaW4udjEuUGx1Z2luSG9zdFNlcnZpY2VBZGRTZXNzaW9uU3RyZWFtUmVxdWVzdBo9LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZUFkZFNlc3Npb25TdHJlYW1SZXNwb25zZRKYAQoTUmVtb3ZlU2Vzc2lvblN0cmVhbRI/LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZVJlbW92ZVNlc3Npb25TdHJlYW1SZXF1ZXN0GkAuaG9sb211c2gucGx1Z2luLnYxLlBsdWdpbkhvc3RTZXJ2aWNlUmVtb3ZlU2Vzc2lvblN0cmVhbVJlc3BvbnNlEnoKCUpvaW5Gb2N1cxI1LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZUpvaW5Gb2N1c1JlcXVlc3QaNi5ob2xvbXVzaC5wbHVnaW4udjEuUGx1Z2luSG9zdFNlcnZpY2VKb2luRm9jdXNSZXNwb25zZRJ9CgpMZWF2ZUZvY3VzEjYuaG9sb211c2gucGx1Z2luLnYxLlBsdWdpbkhvc3RTZXJ2aWNlTGVhdmVGb2N1c1JlcXVlc3QaNy5ob2xvbXVzaC5wbHVnaW4udjEuUGx1Z2luSG9zdFNlcnZpY2VMZWF2ZUZvY3VzUmVzcG9uc2USlQEKEkxlYXZlRm9jdXNCeVRhcmdldBI+LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZUxlYXZlRm9jdXNCeVRhcmdldFJlcXVlc3QaPy5ob2xvbXVzaC5wbHVnaW4udjEuUGx1Z2luSG9zdFNlcnZpY2VMZWF2ZUZvY3VzQnlUYXJnZXRSZXNwb25zZRKDAQoMUHJlc2VudEZvY3VzEjguaG9sb211c2gucGx1Z2luLnYxLlBsdWdpbkhvc3RTZXJ2aWNlUHJlc2VudEZvY3VzUmVxdWVzdBo5LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZVByZXNlbnRGb2N1c1Jlc3BvbnNlEpUBChJRdWVyeVN0cmVhbUhpc3RvcnkSPi5ob2xvbXVzaC5wbHVnaW4udjEuUGx1Z2luSG9zdFNlcnZpY2VRdWVyeVN0cmVhbUhpc3RvcnlSZXF1ZXN0Gj8uaG9sb211c2gucGx1Z2luLnYxLlBsdWdpbkhvc3RTZXJ2aWNlUXVlcnlTdHJlYW1IaXN0b3J5UmVzcG9uc2USdgoTRGVjcnlwdE93bkF1ZGl0Um93cxIuLmhvbG9tdXNoLnBsdWdpbi52MS5EZWNyeXB0T3duQXVkaXRSb3dzUmVxdWVzdBovLmhvbG9tdXNoLnBsdWdpbi52MS5EZWNyeXB0T3duQXVkaXRSb3dzUmVzcG9uc2USjwEKEFJlcXVlc3RFbWl0VG9rZW4SPC5ob2xvbXVzaC5wbHVnaW4udjEuUGx1Z2luSG9zdFNlcnZpY2VSZXF1ZXN0RW1pdFRva2VuUmVxdWVzdBo9LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZVJlcXVlc3RFbWl0VG9rZW5SZXNwb25zZRKVAQoSU2V0Q29ubmVjdGlvbkZvY3VzEj4uaG9sb211c2gucGx1Z2luLnYxLlBsdWdpbkhvc3RTZXJ2aWNlU2V0Q29ubmVjdGlvbkZvY3VzUmVxdWVzdBo/LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZVNldENvbm5lY3Rpb25Gb2N1c1Jlc3BvbnNlEpUBChJHZXRDb25uZWN0aW9uRm9jdXMSPi5ob2xvbXVzaC5wbHVnaW4udjEuUGx1Z2luSG9zdFNlcnZpY2VHZXRDb25uZWN0aW9uRm9jdXNSZXF1ZXN0Gj8uaG9sb211c2gucGx1Z2luLnYxLlBsdWdpbkhvc3RTZXJ2aWNlR2V0Q29ubmVjdGlvbkZvY3VzUmVzcG9uc2USjAEKD0F1dG9Gb2N1c09uSm9pbhI7LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZUF1dG9Gb2N1c09uSm9pblJlcXVlc3QaPC5ob2xvbXVzaC5wbHVnaW4udjEuUGx1Z2luSG9zdFNlcnZpY2VBdXRvRm9jdXNPbkpvaW5SZXNwb25zZRKPAQoQSXNBbnlDb25uRm9jdXNlZBI8LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZUlzQW55Q29ubkZvY3VzZWRSZXF1ZXN0Gj0uaG9sb211c2gucGx1Z2luLnYxLlBsdWdpbkhvc3RTZXJ2aWNlSXNBbnlDb25uRm9jdXNlZFJlc3BvbnNlEncKCEV2YWx1YXRlEjQuaG9sb211c2gucGx1Z2luLnYxLlBsdWdpbkhvc3RTZXJ2aWNlRXZhbHVhdGVSZXF1ZXN0GjUuaG9sb211c2gucGx1Z2luLnYxLlBsdWdpbkhvc3RTZXJ2aWNlRXZhbHVhdGVSZXNwb25zZRKDAQoMTGlzdENvbW1hbmRzEjguaG9sb211c2gucGx1Z2luLnYxLlBsdWdpbkhvc3RTZXJ2aWNlTGlzdENvbW1hbmRzUmVxdWVzdBo5LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZUxpc3RDb21tYW5kc1Jlc3BvbnNlEokBCg5HZXRDb21tYW5kSGVscBI6LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZUdldENvbW1hbmRIZWxwUmVxdWVzdBo7LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZUdldENvbW1hbmRIZWxwUmVzcG9uc2USfQoKR2V0U2V0dGluZxI2LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZUdldFNldHRpbmdSZXF1ZXN0GjcuaG9sb211c2gucGx1Z2luLnYxLlBsdWdpbkhvc3RTZXJ2aWNlR2V0U2V0dGluZ1Jlc3BvbnNlEn0KClNldFNldHRpbmcSNi5ob2xvbXVzaC5wbHVnaW4udjEuUGx1Z2luSG9zdFNlcnZpY2VTZXRTZXR0aW5nUmVxdWVzdBo3LmhvbG9tdXNoLnBsdWdpbi52MS5QbHVnaW5Ib3N0U2VydmljZVNldFNldHRpbmdSZXNwb25zZUJEWkJnaXRodWIuY29tL2hvbG9tdXNoL2hvbG9tdXNoL3BrZy9wcm90by9ob2xvbXVzaC9wbHVnaW4vdjE7cGx1Z2ludjFiBnByb3RvMw", [file_buf_validate_validate, file_holomush_plugin_v1_audit]);
+  fileDesc("Ch9ob2xvbXVzaC9wbHVnaW4vdjEvcGx1Z2luLnByb3RvEhJob2xvbXVzaC5wbHVnaW4udjEiuAIKDVNlcnZpY2VDb25maWcSGQoRY29ubmVjdGlvbl9zdHJpbmcYASABKAkSUgoRcmVxdWlyZWRfc2VydmljZXMYAiADKAsyNy5ob2xvbXVzaC5wbHVnaW4udjEuU2VydmljZUNvbmZpZy5SZXF1aXJlZFNlcnZpY2VzRW50cnkSSgoNcGx1Z2luX2NvbmZpZxgDIAMoCzIzLmhvbG9tdXNoLnBsdWdpbi52MS5TZXJ2aWNlQ29uZmlnLlBsdWdpbkNvbmZpZ0VudHJ5GjcKFVJlcXVpcmVkU2VydmljZXNFbnRyeRILCgNrZXkYASABKAkSDQoFdmFsdWUYAiABKAk6AjgBGjMKEVBsdWdpbkNvbmZpZ0VudHJ5EgsKA2tleRgBIAEoCRINCgV2YWx1ZRgCIAEoCToCOAEiQAoLSW5pdFJlcXVlc3QSMQoGY29uZmlnGAEgASgLMiEuaG9sb211c2gucGx1Z2luLnYxLlNlcnZpY2VDb25maWciSAoMSW5pdFJlc3BvbnNlEhkKEXByb3ZpZGVkX3NlcnZpY2VzGAEgAygJEh0KFXJlZ2lzdGVyZWRfZW1pdF90eXBlcxgCIAMoCSLDAQoFRXZlbnQSEwoCaWQYASABKAlCB7pIBHICEAESFwoGc3RyZWFtGAIgASgJQge6SARyAhABEhUKBHR5cGUYAyABKAlCB7pIBHICEAESEQoJdGltZXN0YW1wGAQgASgDEhsKCmFjdG9yX2tpbmQYBSABKAlCB7pIBHICEAESGQoIYWN0b3JfaWQYBiABKAlCB7pIBHICEAESGgoHcGF5bG9hZBgHIAEoCUIJukgGcgQYgIAEEg4KBmN1cnNvchgIIAEoDCJqCglFbWl0RXZlbnQSFwoGc3RyZWFtGAEgASgJQge6SARyAhABEhUKBHR5cGUYAiABKAlCB7pIBHICEAESGgoHcGF5bG9hZBgDIAEoCUIJukgGcgQYgIAEEhEKCXNlbnNpdGl2ZRgEIAEoCCI+ChJIYW5kbGVFdmVudFJlcXVlc3QSKAoFZXZlbnQYASABKAsyGS5ob2xvbXVzaC5wbHVnaW4udjEuRXZlbnQiSQoTSGFuZGxlRXZlbnRSZXNwb25zZRIyCgtlbWl0X2V2ZW50cxgBIAMoCzIdLmhvbG9tdXNoLnBsdWdpbi52MS5FbWl0RXZlbnQijQIKDkNvbW1hbmRSZXF1ZXN0EhgKB2NvbW1hbmQYASABKAlCB7pIBHICEAESFgoEYXJncxgCIAEoCUIIukgFcgMYgEASGwoJcmF3X2lucHV0GAMgASgJQgi6SAVyAxiAQBIdCgxjaGFyYWN0ZXJfaWQYBCABKAlCB7pIBHICEAESHwoOY2hhcmFjdGVyX25hbWUYBSABKAlCB7pIBHICEAESHAoLbG9jYXRpb25faWQYBiABKAlCB7pIBHICEAESGwoKc2Vzc2lvbl9pZBgHIAEoCUIHukgEcgIQARIaCglwbGF5ZXJfaWQYCCABKAlCB7pIBHICEAESFQoNY29ubmVjdGlvbl9pZBgJIAEoCSLJAQoPQ29tbWFuZFJlc3BvbnNlEjEKBnN0YXR1cxgBIAEoDjIhLmhvbG9tdXNoLnBsdWdpbi52MS5Db21tYW5kU3RhdHVzEhgKBm91dHB1dBgCIAEoCUIIukgFcgMYgEASLQoGZXZlbnRzGAMgAygLMh0uaG9sb211c2gucGx1Z2luLnYxLkVtaXRFdmVudBI6CgthdWRpdF9oaW50cxgEIAMoCzIlLmhvbG9tdXNoLnBsdWdpbi52MS5BdWRpdERlY2lzaW9uSGludCLMAgoRQXVkaXREZWNpc2lvbkhpbnQSFgoCaWQYASABKAlCCrpIB3IFEAEYgAESFgoEbmFtZRgCIAEoCUIIukgFcgMYgAISGQoHbWVzc2FnZRgDIAEoCUIIukgFcgMYgAgSLwoGZWZmZWN0GAQgASgOMh8uaG9sb211c2gucGx1Z2luLnYxLkF1ZGl0RWZmZWN0EiEKEGFjdGlvbl9xdWFsaWZpZXIYBSABKAlCB7pIBHICGEASGgoIcmVzb3VyY2UYBiABKAlCCLpIBXIDGIACEkkKCmF0dHJpYnV0ZXMYByADKAsyNS5ob2xvbXVzaC5wbHVnaW4udjEuQXVkaXREZWNpc2lvbkhpbnQuQXR0cmlidXRlc0VudHJ5GjEKD0F0dHJpYnV0ZXNFbnRyeRILCgNrZXkYASABKAkSDQoFdmFsdWUYAiABKAk6AjgBIksKFEhhbmRsZUNvbW1hbmRSZXF1ZXN0EjMKB2NvbW1hbmQYASABKAsyIi5ob2xvbXVzaC5wbHVnaW4udjEuQ29tbWFuZFJlcXVlc3QiTgoVSGFuZGxlQ29tbWFuZFJlc3BvbnNlEjUKCHJlc3BvbnNlGAEgASgLMiMuaG9sb211c2gucGx1Z2luLnYxLkNvbW1hbmRSZXNwb25zZSJZChpRdWVyeVNlc3Npb25TdHJlYW1zUmVxdWVzdBIUCgxjaGFyYWN0ZXJfaWQYASABKAkSEQoJcGxheWVyX2lkGAIgASgJEhIKCnNlc3Npb25faWQYAyABKAkiPQobUXVlcnlTZXNzaW9uU3RyZWFtc1Jlc3BvbnNlEg8KB3N0cmVhbXMYASADKAkSDQoFZXJyb3IYAiABKAkqlgEKDUNvbW1hbmRTdGF0dXMSHgoaQ09NTUFORF9TVEFUVVNfVU5TUEVDSUZJRUQQABIVChFDT01NQU5EX1NUQVRVU19PSxABEhgKFENPTU1BTkRfU1RBVFVTX0VSUk9SEAISGgoWQ09NTUFORF9TVEFUVVNfRkFJTFVSRRADEhgKFENPTU1BTkRfU1RBVFVTX0ZBVEFMEAQqWgoLQXVkaXRFZmZlY3QSHAoYQVVESVRfRUZGRUNUX1VOU1BFQ0lGSUVEEAASFQoRQVVESVRfRUZGRUNUX0RFTlkQARIWChJBVURJVF9FRkZFQ1RfQUxMT1cQAip8CgxTZXR0aW5nU2NvcGUSHQoZU0VUVElOR19TQ09QRV9VTlNQRUNJRklFRBAAEhYKElNFVFRJTkdfU0NPUEVfR0FNRRABEhgKFFNFVFRJTkdfU0NPUEVfUExBWUVSEAISGwoXU0VUVElOR19TQ09QRV9DSEFSQUNURVIQAzKYAwoNUGx1Z2luU2VydmljZRJJCgRJbml0Eh8uaG9sb211c2gucGx1Z2luLnYxLkluaXRSZXF1ZXN0GiAuaG9sb211c2gucGx1Z2luLnYxLkluaXRSZXNwb25zZRJeCgtIYW5kbGVFdmVudBImLmhvbG9tdXNoLnBsdWdpbi52MS5IYW5kbGVFdmVudFJlcXVlc3QaJy5ob2xvbXVzaC5wbHVnaW4udjEuSGFuZGxlRXZlbnRSZXNwb25zZRJkCg1IYW5kbGVDb21tYW5kEiguaG9sb211c2gucGx1Z2luLnYxLkhhbmRsZUNvbW1hbmRSZXF1ZXN0GikuaG9sb211c2gucGx1Z2luLnYxLkhhbmRsZUNvbW1hbmRSZXNwb25zZRJ2ChNRdWVyeVNlc3Npb25TdHJlYW1zEi4uaG9sb211c2gucGx1Z2luLnYxLlF1ZXJ5U2Vzc2lvblN0cmVhbXNSZXF1ZXN0Gi8uaG9sb211c2gucGx1Z2luLnYxLlF1ZXJ5U2Vzc2lvblN0cmVhbXNSZXNwb25zZUJEWkJnaXRodWIuY29tL2hvbG9tdXNoL2hvbG9tdXNoL3BrZy9wcm90by9ob2xvbXVzaC9wbHVnaW4vdjE7cGx1Z2ludjFiBnByb3RvMw", [file_buf_validate_validate]);
 
 /**
  * ServiceConfig carries the host→plugin initialization payload delivered in
@@ -181,10 +179,10 @@ export type Event = Message<"holomush.plugin.v1.Event"> & {
   payload: string;
 
   /**
-   * cursor is the opaque pagination token for this event. Pass as
-   * PluginHostServiceQueryStreamHistoryRequest.cursor on the next call
-   * to page backward from this position. Empty on events received via
-   * delivery (not history). Treat as an opaque blob.
+   * cursor is the opaque pagination token for this event. Pass as the
+   * QueryStreamHistory request cursor on the next call to page backward from
+   * this position. Empty on events received via delivery (not history). Treat
+   * as an opaque blob.
    *
    * @generated from field: bytes cursor = 8;
    */
@@ -544,372 +542,6 @@ export const HandleCommandResponseSchema: GenMessage<HandleCommandResponse> = /*
   messageDesc(file_holomush_plugin_v1_plugin, 11);
 
 /**
- * PluginHostServiceEmitEventRequest is the wire form of a plugin emit. The
- * caller's identity is NOT on this message — it is recovered host-side from the
- * x-holomush-emit-token header (see PluginHostService.EmitEvent).
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceEmitEventRequest
- */
-export type PluginHostServiceEmitEventRequest = Message<"holomush.plugin.v1.PluginHostServiceEmitEventRequest"> & {
-  /**
-   * Target stream (legacy "prefix:id" form); its namespace must be declared in
-   * the manifest's emits list or the fence rejects the emit.
-   *
-   * @generated from field: string stream = 1;
-   */
-  stream: string;
-
-  /**
-   * Event-type discriminator for the emitted event.
-   *
-   * @generated from field: string event_type = 2;
-   */
-  eventType: string;
-
-  /**
-   * Raw event payload bytes (validated as JSON at the fence).
-   *
-   * @generated from field: bytes payload = 3;
-   */
-  payload: Uint8Array;
-
-  /**
-   * sensitive declares per-event sensitivity at emit time.
-   * Phase 3a's host-side fence at internal/plugin/event_emitter.go::Emit
-   * validates this against the plugin manifest's declared sensitivity:
-   *   - manifest sensitivity=never:  sensitive=true rejected (INV-PLUGIN-29).
-   *   - manifest sensitivity=may:    sensitive=true/false honored.
-   *   - manifest sensitivity=always: sensitive=false rejected (INV-PLUGIN-30).
-   * Default false (proto3 zero) for older plugins compiled before this
-   * field existed — matching pre-Phase-3d behavior.
-   *
-   * @generated from field: bool sensitive = 4;
-   */
-  sensitive: boolean;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceEmitEventRequest.
- * Use `create(PluginHostServiceEmitEventRequestSchema)` to create a new message.
- */
-export const PluginHostServiceEmitEventRequestSchema: GenMessage<PluginHostServiceEmitEventRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 12);
-
-/**
- * PluginHostServiceEmitEventResponse is the empty acknowledgement that an emit
- * passed the fence and was published.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceEmitEventResponse
- */
-export type PluginHostServiceEmitEventResponse = Message<"holomush.plugin.v1.PluginHostServiceEmitEventResponse"> & {
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceEmitEventResponse.
- * Use `create(PluginHostServiceEmitEventResponseSchema)` to create a new message.
- */
-export const PluginHostServiceEmitEventResponseSchema: GenMessage<PluginHostServiceEmitEventResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 13);
-
-/**
- * PluginHostServiceLogRequest is the (currently unserved, holomush-l6std)
- * request to write one plugin log line through the host logger.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceLogRequest
- */
-export type PluginHostServiceLogRequest = Message<"holomush.plugin.v1.PluginHostServiceLogRequest"> & {
-  /**
-   * Log severity (e.g. "info", "warn", "error").
-   *
-   * @generated from field: string level = 1;
-   */
-  level: string;
-
-  /**
-   * Log message body (max 8 KiB).
-   *
-   * @generated from field: string message = 2;
-   */
-  message: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceLogRequest.
- * Use `create(PluginHostServiceLogRequestSchema)` to create a new message.
- */
-export const PluginHostServiceLogRequestSchema: GenMessage<PluginHostServiceLogRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 14);
-
-/**
- * PluginHostServiceLogResponse is the empty ack for Log (unserved,
- * holomush-l6std).
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceLogResponse
- */
-export type PluginHostServiceLogResponse = Message<"holomush.plugin.v1.PluginHostServiceLogResponse"> & {
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceLogResponse.
- * Use `create(PluginHostServiceLogResponseSchema)` to create a new message.
- */
-export const PluginHostServiceLogResponseSchema: GenMessage<PluginHostServiceLogResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 15);
-
-/**
- * PluginHostServiceKVGetRequest is the (currently unserved, holomush-l6std)
- * request to read a key from a plugin's KV namespace.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceKVGetRequest
- */
-export type PluginHostServiceKVGetRequest = Message<"holomush.plugin.v1.PluginHostServiceKVGetRequest"> & {
-  /**
-   * Owning plugin's name — the KV namespace key.
-   *
-   * @generated from field: string plugin_name = 1;
-   */
-  pluginName: string;
-
-  /**
-   * Key to read within that namespace.
-   *
-   * @generated from field: string key = 2;
-   */
-  key: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceKVGetRequest.
- * Use `create(PluginHostServiceKVGetRequestSchema)` to create a new message.
- */
-export const PluginHostServiceKVGetRequestSchema: GenMessage<PluginHostServiceKVGetRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 16);
-
-/**
- * PluginHostServiceKVGetResponse returns a KV lookup result (unserved,
- * holomush-l6std).
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceKVGetResponse
- */
-export type PluginHostServiceKVGetResponse = Message<"holomush.plugin.v1.PluginHostServiceKVGetResponse"> & {
-  /**
-   * The stored value, or empty when not found.
-   *
-   * @generated from field: string value = 1;
-   */
-  value: string;
-
-  /**
-   * Whether the key existed; distinguishes a stored empty value from a miss.
-   *
-   * @generated from field: bool found = 2;
-   */
-  found: boolean;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceKVGetResponse.
- * Use `create(PluginHostServiceKVGetResponseSchema)` to create a new message.
- */
-export const PluginHostServiceKVGetResponseSchema: GenMessage<PluginHostServiceKVGetResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 17);
-
-/**
- * PluginHostServiceKVSetRequest is the (currently unserved, holomush-l6std)
- * request to write a key in a plugin's KV namespace.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceKVSetRequest
- */
-export type PluginHostServiceKVSetRequest = Message<"holomush.plugin.v1.PluginHostServiceKVSetRequest"> & {
-  /**
-   * Owning plugin's name — the KV namespace key.
-   *
-   * @generated from field: string plugin_name = 1;
-   */
-  pluginName: string;
-
-  /**
-   * Key to write within that namespace.
-   *
-   * @generated from field: string key = 2;
-   */
-  key: string;
-
-  /**
-   * Value to store under the key.
-   *
-   * @generated from field: string value = 3;
-   */
-  value: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceKVSetRequest.
- * Use `create(PluginHostServiceKVSetRequestSchema)` to create a new message.
- */
-export const PluginHostServiceKVSetRequestSchema: GenMessage<PluginHostServiceKVSetRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 18);
-
-/**
- * PluginHostServiceKVSetResponse is the empty ack for KVSet (unserved,
- * holomush-l6std).
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceKVSetResponse
- */
-export type PluginHostServiceKVSetResponse = Message<"holomush.plugin.v1.PluginHostServiceKVSetResponse"> & {
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceKVSetResponse.
- * Use `create(PluginHostServiceKVSetResponseSchema)` to create a new message.
- */
-export const PluginHostServiceKVSetResponseSchema: GenMessage<PluginHostServiceKVSetResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 19);
-
-/**
- * PluginHostServiceKVDeleteRequest is the (currently unserved, holomush-l6std)
- * request to delete a key from a plugin's KV namespace.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceKVDeleteRequest
- */
-export type PluginHostServiceKVDeleteRequest = Message<"holomush.plugin.v1.PluginHostServiceKVDeleteRequest"> & {
-  /**
-   * Owning plugin's name — the KV namespace key.
-   *
-   * @generated from field: string plugin_name = 1;
-   */
-  pluginName: string;
-
-  /**
-   * Key to delete within that namespace.
-   *
-   * @generated from field: string key = 2;
-   */
-  key: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceKVDeleteRequest.
- * Use `create(PluginHostServiceKVDeleteRequestSchema)` to create a new message.
- */
-export const PluginHostServiceKVDeleteRequestSchema: GenMessage<PluginHostServiceKVDeleteRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 20);
-
-/**
- * PluginHostServiceKVDeleteResponse is the empty ack for KVDelete (unserved,
- * holomush-l6std).
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceKVDeleteResponse
- */
-export type PluginHostServiceKVDeleteResponse = Message<"holomush.plugin.v1.PluginHostServiceKVDeleteResponse"> & {
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceKVDeleteResponse.
- * Use `create(PluginHostServiceKVDeleteResponseSchema)` to create a new message.
- */
-export const PluginHostServiceKVDeleteResponseSchema: GenMessage<PluginHostServiceKVDeleteResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 21);
-
-/**
- * PluginHostServiceAddSessionStreamRequest is the (currently unserved,
- * holomush-l6std) request to subscribe an active session to one more stream.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceAddSessionStreamRequest
- */
-export type PluginHostServiceAddSessionStreamRequest = Message<"holomush.plugin.v1.PluginHostServiceAddSessionStreamRequest"> & {
-  /**
-   * Active session identifier.
-   *
-   * @generated from field: string session_id = 1;
-   */
-  sessionId: string;
-
-  /**
-   * Stream name to subscribe to (format: "prefix:id").
-   *
-   * @generated from field: string stream = 2;
-   */
-  stream: string;
-
-  /**
-   * replay_mode controls initial replay. Optional; defaults to
-   * FROM_CURSOR if unspecified for backwards compatibility.
-   *
-   * @generated from field: holomush.plugin.v1.StreamReplayMode replay_mode = 3;
-   */
-  replayMode: StreamReplayMode;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceAddSessionStreamRequest.
- * Use `create(PluginHostServiceAddSessionStreamRequestSchema)` to create a new message.
- */
-export const PluginHostServiceAddSessionStreamRequestSchema: GenMessage<PluginHostServiceAddSessionStreamRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 22);
-
-/**
- * PluginHostServiceAddSessionStreamResponse is the empty ack for
- * AddSessionStream (unserved, holomush-l6std).
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceAddSessionStreamResponse
- */
-export type PluginHostServiceAddSessionStreamResponse = Message<"holomush.plugin.v1.PluginHostServiceAddSessionStreamResponse"> & {
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceAddSessionStreamResponse.
- * Use `create(PluginHostServiceAddSessionStreamResponseSchema)` to create a new message.
- */
-export const PluginHostServiceAddSessionStreamResponseSchema: GenMessage<PluginHostServiceAddSessionStreamResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 23);
-
-/**
- * PluginHostServiceRemoveSessionStreamRequest is the (currently unserved,
- * holomush-l6std) request to unsubscribe an active session from a stream.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceRemoveSessionStreamRequest
- */
-export type PluginHostServiceRemoveSessionStreamRequest = Message<"holomush.plugin.v1.PluginHostServiceRemoveSessionStreamRequest"> & {
-  /**
-   * Active session identifier.
-   *
-   * @generated from field: string session_id = 1;
-   */
-  sessionId: string;
-
-  /**
-   * Stream name to unsubscribe from (format: "prefix:id").
-   *
-   * @generated from field: string stream = 2;
-   */
-  stream: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceRemoveSessionStreamRequest.
- * Use `create(PluginHostServiceRemoveSessionStreamRequestSchema)` to create a new message.
- */
-export const PluginHostServiceRemoveSessionStreamRequestSchema: GenMessage<PluginHostServiceRemoveSessionStreamRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 24);
-
-/**
- * PluginHostServiceRemoveSessionStreamResponse is the empty ack for
- * RemoveSessionStream (unserved, holomush-l6std).
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceRemoveSessionStreamResponse
- */
-export type PluginHostServiceRemoveSessionStreamResponse = Message<"holomush.plugin.v1.PluginHostServiceRemoveSessionStreamResponse"> & {
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceRemoveSessionStreamResponse.
- * Use `create(PluginHostServiceRemoveSessionStreamResponseSchema)` to create a new message.
- */
-export const PluginHostServiceRemoveSessionStreamResponseSchema: GenMessage<PluginHostServiceRemoveSessionStreamResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 25);
-
-/**
  * QuerySessionStreamsRequest provides session context for stream contribution.
  *
  * @generated from message holomush.plugin.v1.QuerySessionStreamsRequest
@@ -942,7 +574,7 @@ export type QuerySessionStreamsRequest = Message<"holomush.plugin.v1.QuerySessio
  * Use `create(QuerySessionStreamsRequestSchema)` to create a new message.
  */
 export const QuerySessionStreamsRequestSchema: GenMessage<QuerySessionStreamsRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 26);
+  messageDesc(file_holomush_plugin_v1_plugin, 12);
 
 /**
  * QuerySessionStreamsResponse returns stream names and an optional error.
@@ -970,976 +602,7 @@ export type QuerySessionStreamsResponse = Message<"holomush.plugin.v1.QuerySessi
  * Use `create(QuerySessionStreamsResponseSchema)` to create a new message.
  */
 export const QuerySessionStreamsResponseSchema: GenMessage<QuerySessionStreamsResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 27);
-
-/**
- * FocusKey identifies a focus membership within a session. A session's
- * focus memberships are unique by (kind, target_id) pair.
- *
- * @generated from message holomush.plugin.v1.FocusKey
- */
-export type FocusKey = Message<"holomush.plugin.v1.FocusKey"> & {
-  /**
-   * Which kind of focused context this key names.
-   *
-   * @generated from field: holomush.plugin.v1.FocusKind kind = 1;
-   */
-  kind: FocusKind;
-
-  /**
-   * ULID of the focused target (e.g. the scene id) within that kind.
-   *
-   * @generated from field: string target_id = 2;
-   */
-  targetId: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.FocusKey.
- * Use `create(FocusKeySchema)` to create a new message.
- */
-export const FocusKeySchema: GenMessage<FocusKey> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 28);
-
-/**
- * PluginHostServiceJoinFocusRequest names the session and the focus target to
- * add a membership for.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceJoinFocusRequest
- */
-export type PluginHostServiceJoinFocusRequest = Message<"holomush.plugin.v1.PluginHostServiceJoinFocusRequest"> & {
-  /**
-   * Session to add the focus membership to.
-   *
-   * @generated from field: string session_id = 1;
-   */
-  sessionId: string;
-
-  /**
-   * The (kind, target_id) membership to add.
-   *
-   * @generated from field: holomush.plugin.v1.FocusKey target = 2;
-   */
-  target?: FocusKey | undefined;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceJoinFocusRequest.
- * Use `create(PluginHostServiceJoinFocusRequestSchema)` to create a new message.
- */
-export const PluginHostServiceJoinFocusRequestSchema: GenMessage<PluginHostServiceJoinFocusRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 29);
-
-/**
- * PluginHostServiceJoinFocusResponse is the empty ack that the membership was
- * added.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceJoinFocusResponse
- */
-export type PluginHostServiceJoinFocusResponse = Message<"holomush.plugin.v1.PluginHostServiceJoinFocusResponse"> & {
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceJoinFocusResponse.
- * Use `create(PluginHostServiceJoinFocusResponseSchema)` to create a new message.
- */
-export const PluginHostServiceJoinFocusResponseSchema: GenMessage<PluginHostServiceJoinFocusResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 30);
-
-/**
- * PluginHostServiceLeaveFocusRequest names the session and the focus target to
- * remove a membership for (idempotent on non-member).
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceLeaveFocusRequest
- */
-export type PluginHostServiceLeaveFocusRequest = Message<"holomush.plugin.v1.PluginHostServiceLeaveFocusRequest"> & {
-  /**
-   * Session to remove the focus membership from.
-   *
-   * @generated from field: string session_id = 1;
-   */
-  sessionId: string;
-
-  /**
-   * The (kind, target_id) membership to remove.
-   *
-   * @generated from field: holomush.plugin.v1.FocusKey target = 2;
-   */
-  target?: FocusKey | undefined;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceLeaveFocusRequest.
- * Use `create(PluginHostServiceLeaveFocusRequestSchema)` to create a new message.
- */
-export const PluginHostServiceLeaveFocusRequestSchema: GenMessage<PluginHostServiceLeaveFocusRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 31);
-
-/**
- * PluginHostServiceLeaveFocusResponse is the empty ack for LeaveFocus.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceLeaveFocusResponse
- */
-export type PluginHostServiceLeaveFocusResponse = Message<"holomush.plugin.v1.PluginHostServiceLeaveFocusResponse"> & {
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceLeaveFocusResponse.
- * Use `create(PluginHostServiceLeaveFocusResponseSchema)` to create a new message.
- */
-export const PluginHostServiceLeaveFocusResponseSchema: GenMessage<PluginHostServiceLeaveFocusResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 32);
-
-/**
- * PluginHostServiceLeaveFocusByTargetRequest names a focus target to remove
- * from every session that holds it (cross-session fan-out).
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceLeaveFocusByTargetRequest
- */
-export type PluginHostServiceLeaveFocusByTargetRequest = Message<"holomush.plugin.v1.PluginHostServiceLeaveFocusByTargetRequest"> & {
-  /**
-   * The (kind, target_id) membership to sweep out of all holding sessions.
-   *
-   * @generated from field: holomush.plugin.v1.FocusKey target = 1;
-   */
-  target?: FocusKey | undefined;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceLeaveFocusByTargetRequest.
- * Use `create(PluginHostServiceLeaveFocusByTargetRequestSchema)` to create a new message.
- */
-export const PluginHostServiceLeaveFocusByTargetRequestSchema: GenMessage<PluginHostServiceLeaveFocusByTargetRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 33);
-
-/**
- * PluginHostServiceLeaveFocusByTargetResponse reports the aggregate result of a
- * cross-session leave sweep.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceLeaveFocusByTargetResponse
- */
-export type PluginHostServiceLeaveFocusByTargetResponse = Message<"holomush.plugin.v1.PluginHostServiceLeaveFocusByTargetResponse"> & {
-  /**
-   * Number of sessions successfully left. Zero is a valid result
-   * (target had no members or every member was already non-a-member
-   * — per-session idempotent no-ops count as successes). Callers
-   * comparing succeeded + len(failed_session_ids) against total_scanned
-   * can distinguish total, partial, and empty-sweep outcomes without
-   * parsing any error string.
-   *
-   * @generated from field: int32 succeeded = 1;
-   */
-  succeeded: number;
-
-  /**
-   * Number of non-expired sessions the sweep scanned. Always
-   * >= succeeded + len(failed_session_ids).
-   *
-   * @generated from field: int32 total_scanned = 2;
-   */
-  totalScanned: number;
-
-  /**
-   * Session IDs for which the per-session leave failed. Empty means
-   * every scanned session succeeded (idempotent no-ops included).
-   * Per-session error details are not serialized; callers should
-   * treat these IDs as the authoritative partial-failure signal and
-   * re-issue LeaveFocus against them if retry is desired. The RPC
-   * error is reserved for enumeration/list failures (e.g., the
-   * session store could not list members).
-   *
-   * @generated from field: repeated string failed_session_ids = 3;
-   */
-  failedSessionIds: string[];
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceLeaveFocusByTargetResponse.
- * Use `create(PluginHostServiceLeaveFocusByTargetResponseSchema)` to create a new message.
- */
-export const PluginHostServiceLeaveFocusByTargetResponseSchema: GenMessage<PluginHostServiceLeaveFocusByTargetResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 34);
-
-/**
- * PluginHostServicePresentFocusRequest names the session and the existing
- * membership to set as its PresentingFocus.
- *
- * @generated from message holomush.plugin.v1.PluginHostServicePresentFocusRequest
- */
-export type PluginHostServicePresentFocusRequest = Message<"holomush.plugin.v1.PluginHostServicePresentFocusRequest"> & {
-  /**
-   * Session whose PresentingFocus pointer is being set.
-   *
-   * @generated from field: string session_id = 1;
-   */
-  sessionId: string;
-
-  /**
-   * Existing membership to present; validated against FocusMemberships.
-   *
-   * @generated from field: holomush.plugin.v1.FocusKey target = 2;
-   */
-  target?: FocusKey | undefined;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServicePresentFocusRequest.
- * Use `create(PluginHostServicePresentFocusRequestSchema)` to create a new message.
- */
-export const PluginHostServicePresentFocusRequestSchema: GenMessage<PluginHostServicePresentFocusRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 35);
-
-/**
- * PluginHostServicePresentFocusResponse is the empty ack for PresentFocus.
- *
- * @generated from message holomush.plugin.v1.PluginHostServicePresentFocusResponse
- */
-export type PluginHostServicePresentFocusResponse = Message<"holomush.plugin.v1.PluginHostServicePresentFocusResponse"> & {
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServicePresentFocusResponse.
- * Use `create(PluginHostServicePresentFocusResponseSchema)` to create a new message.
- */
-export const PluginHostServicePresentFocusResponseSchema: GenMessage<PluginHostServicePresentFocusResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 36);
-
-/**
- * PluginHostServiceQueryStreamHistoryRequest selects a backward-paginated tail
- * of a stream for plugin-side display.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceQueryStreamHistoryRequest
- */
-export type PluginHostServiceQueryStreamHistoryRequest = Message<"holomush.plugin.v1.PluginHostServiceQueryStreamHistoryRequest"> & {
-  /**
-   * Stream to read history from (legacy "prefix:id" form).
-   *
-   * @generated from field: string stream = 1;
-   */
-  stream: string;
-
-  /**
-   * Page size; negative is rejected, values above 500 are clamped to 500.
-   *
-   * @generated from field: int32 count = 2;
-   */
-  count: number;
-
-  /**
-   * Epoch milliseconds. Events before this time are excluded. 0 means no lower bound.
-   *
-   * @generated from field: int64 not_before_ms = 3;
-   */
-  notBeforeMs: bigint;
-
-  /**
-   * cursor is the opaque pagination cursor from a previous response.
-   * Events older than the cursor position are returned. Empty = start from latest.
-   *
-   * @generated from field: bytes cursor = 4;
-   */
-  cursor: Uint8Array;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceQueryStreamHistoryRequest.
- * Use `create(PluginHostServiceQueryStreamHistoryRequestSchema)` to create a new message.
- */
-export const PluginHostServiceQueryStreamHistoryRequestSchema: GenMessage<PluginHostServiceQueryStreamHistoryRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 37);
-
-/**
- * PluginHostServiceQueryStreamHistoryResponse returns one history page plus the
- * cursor for the next (older) page.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceQueryStreamHistoryResponse
- */
-export type PluginHostServiceQueryStreamHistoryResponse = Message<"holomush.plugin.v1.PluginHostServiceQueryStreamHistoryResponse"> & {
-  /**
-   * The page of events in ascending (oldest→newest) order; each carries its own
-   * backward-paging cursor.
-   *
-   * @generated from field: repeated holomush.plugin.v1.Event events = 1;
-   */
-  events: Event[];
-
-  /**
-   * next_cursor is the opaque cursor for the next page. Empty if no more pages.
-   *
-   * @generated from field: bytes next_cursor = 2;
-   */
-  nextCursor: Uint8Array;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceQueryStreamHistoryResponse.
- * Use `create(PluginHostServiceQueryStreamHistoryResponseSchema)` to create a new message.
- */
-export const PluginHostServiceQueryStreamHistoryResponseSchema: GenMessage<PluginHostServiceQueryStreamHistoryResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 38);
-
-/**
- * PluginHostServiceRequestEmitTokenRequest carries no fields. The host
- * derives the calling plugin's identity from the mTLS-bound server struct.
- * Future evolution: do NOT add actor fields here — that would re-open the
- * G1 forgery surface this RPC is designed to close.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceRequestEmitTokenRequest
- */
-export type PluginHostServiceRequestEmitTokenRequest = Message<"holomush.plugin.v1.PluginHostServiceRequestEmitTokenRequest"> & {
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceRequestEmitTokenRequest.
- * Use `create(PluginHostServiceRequestEmitTokenRequestSchema)` to create a new message.
- */
-export const PluginHostServiceRequestEmitTokenRequestSchema: GenMessage<PluginHostServiceRequestEmitTokenRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 39);
-
-/**
- * PluginHostServiceRequestEmitTokenResponse returns the issued self-token.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceRequestEmitTokenResponse
- */
-export type PluginHostServiceRequestEmitTokenResponse = Message<"holomush.plugin.v1.PluginHostServiceRequestEmitTokenResponse"> & {
-  /**
-   * Opaque self-token. Plugins MUST treat this as opaque; only the host's
-   * emitTokenStore can interpret it. The token is bound to ActorPlugin +
-   * the calling plugin's name and is single-use-friendly (TTL-revoked).
-   *
-   * @generated from field: string token = 1;
-   */
-  token: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceRequestEmitTokenResponse.
- * Use `create(PluginHostServiceRequestEmitTokenResponseSchema)` to create a new message.
- */
-export const PluginHostServiceRequestEmitTokenResponseSchema: GenMessage<PluginHostServiceRequestEmitTokenResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 40);
-
-/**
- * PluginHostServiceSetConnectionFocusRequest selects one connection and the
- * focus to set on it (Phase 5).
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceSetConnectionFocusRequest
- */
-export type PluginHostServiceSetConnectionFocusRequest = Message<"holomush.plugin.v1.PluginHostServiceSetConnectionFocusRequest"> & {
-  /**
-   * ULID bytes of the connection whose focus is being set.
-   *
-   * @generated from field: bytes connection_id = 1;
-   */
-  connectionId: Uint8Array;
-
-  /**
-   * The focus to set; absent (unset) clears the connection's focus.
-   *
-   * @generated from field: optional holomush.plugin.v1.FocusKey focus_key = 2;
-   */
-  focusKey?: FocusKey | undefined;
-
-  /**
-   * is_scene_grid signals that this call originated from a `scene grid`
-   * command — substrate skips the D9 PresentingFocus write per D10.
-   *
-   * @generated from field: bool is_scene_grid = 3;
-   */
-  isSceneGrid: boolean;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceSetConnectionFocusRequest.
- * Use `create(PluginHostServiceSetConnectionFocusRequestSchema)` to create a new message.
- */
-export const PluginHostServiceSetConnectionFocusRequestSchema: GenMessage<PluginHostServiceSetConnectionFocusRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 41);
-
-/**
- * PluginHostServiceSetConnectionFocusResponse echoes the resulting focus.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceSetConnectionFocusResponse
- */
-export type PluginHostServiceSetConnectionFocusResponse = Message<"holomush.plugin.v1.PluginHostServiceSetConnectionFocusResponse"> & {
-  /**
-   * The connection's focus after the write; absent if the connection was
-   * cleared/unfocused.
-   *
-   * @generated from field: optional holomush.plugin.v1.FocusKey focus_key = 1;
-   */
-  focusKey?: FocusKey | undefined;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceSetConnectionFocusResponse.
- * Use `create(PluginHostServiceSetConnectionFocusResponseSchema)` to create a new message.
- */
-export const PluginHostServiceSetConnectionFocusResponseSchema: GenMessage<PluginHostServiceSetConnectionFocusResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 42);
-
-/**
- * PluginHostServiceGetConnectionFocusRequest carries the connection ID whose
- * focus is being read. Read-only counterpart of SetConnectionFocus.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceGetConnectionFocusRequest
- */
-export type PluginHostServiceGetConnectionFocusRequest = Message<"holomush.plugin.v1.PluginHostServiceGetConnectionFocusRequest"> & {
-  /**
-   * ULID bytes of the connection whose focus is being read.
-   *
-   * @generated from field: bytes connection_id = 1;
-   */
-  connectionId: Uint8Array;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceGetConnectionFocusRequest.
- * Use `create(PluginHostServiceGetConnectionFocusRequestSchema)` to create a new message.
- */
-export const PluginHostServiceGetConnectionFocusRequestSchema: GenMessage<PluginHostServiceGetConnectionFocusRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 43);
-
-/**
- * PluginHostServiceGetConnectionFocusResponse returns the connection's current
- * focus key, if any. Absent when the connection is grid-focused or unknown.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceGetConnectionFocusResponse
- */
-export type PluginHostServiceGetConnectionFocusResponse = Message<"holomush.plugin.v1.PluginHostServiceGetConnectionFocusResponse"> & {
-  /**
-   * The connection's focus; absent for grid focus or unknown connection.
-   *
-   * @generated from field: optional holomush.plugin.v1.FocusKey focus_key = 1;
-   */
-  focusKey?: FocusKey | undefined;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceGetConnectionFocusResponse.
- * Use `create(PluginHostServiceGetConnectionFocusResponseSchema)` to create a new message.
- */
-export const PluginHostServiceGetConnectionFocusResponseSchema: GenMessage<PluginHostServiceGetConnectionFocusResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 44);
-
-/**
- * PluginHostServiceAutoFocusOnJoinRequest names the character and scene to
- * fan-out focus across the character's connections.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceAutoFocusOnJoinRequest
- */
-export type PluginHostServiceAutoFocusOnJoinRequest = Message<"holomush.plugin.v1.PluginHostServiceAutoFocusOnJoinRequest"> & {
-  /**
-   * ULID bytes of the character whose connections are being focused.
-   *
-   * @generated from field: bytes character_id = 1;
-   */
-  characterId: Uint8Array;
-
-  /**
-   * ULID bytes of the scene to focus those connections on.
-   *
-   * @generated from field: bytes scene_id = 2;
-   */
-  sceneId: Uint8Array;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceAutoFocusOnJoinRequest.
- * Use `create(PluginHostServiceAutoFocusOnJoinRequestSchema)` to create a new message.
- */
-export const PluginHostServiceAutoFocusOnJoinRequestSchema: GenMessage<PluginHostServiceAutoFocusOnJoinRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 45);
-
-/**
- * PluginHostServiceAutoFocusOnJoinResponse reports per-connection fan-out
- * outcomes.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceAutoFocusOnJoinResponse
- */
-export type PluginHostServiceAutoFocusOnJoinResponse = Message<"holomush.plugin.v1.PluginHostServiceAutoFocusOnJoinResponse"> & {
-  /**
-   * Connection ULIDs newly focused on the scene by this call.
-   *
-   * @generated from field: repeated bytes focused_connection_ids = 1;
-   */
-  focusedConnectionIds: Uint8Array[];
-
-  /**
-   * Total terminal/telnet connections the character had (the fan-out
-   * denominator).
-   *
-   * @generated from field: uint32 total_connection_count = 2;
-   */
-  totalConnectionCount: number;
-
-  /**
-   * Connection ULIDs skipped because they were already explicitly focused
-   * elsewhere (D8).
-   *
-   * @generated from field: repeated bytes skipped_connection_ids = 3;
-   */
-  skippedConnectionIds: Uint8Array[];
-
-  /**
-   * Connections that failed to focus, each with a structured reason.
-   *
-   * @generated from field: repeated holomush.plugin.v1.FocusFailure failed_connection_ids = 4;
-   */
-  failedConnectionIds: FocusFailure[];
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceAutoFocusOnJoinResponse.
- * Use `create(PluginHostServiceAutoFocusOnJoinResponseSchema)` to create a new message.
- */
-export const PluginHostServiceAutoFocusOnJoinResponseSchema: GenMessage<PluginHostServiceAutoFocusOnJoinResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 46);
-
-/**
- * FocusFailure carries the connection_id and reason for an AutoFocusOnJoin failure.
- *
- * @generated from message holomush.plugin.v1.FocusFailure
- */
-export type FocusFailure = Message<"holomush.plugin.v1.FocusFailure"> & {
-  /**
-   * ULID bytes of the connection that failed to focus.
-   *
-   * @generated from field: bytes connection_id = 1;
-   */
-  connectionId: Uint8Array;
-
-  /**
-   * Why the focus attempt failed for that connection.
-   *
-   * @generated from field: holomush.plugin.v1.FocusFailureReason reason = 2;
-   */
-  reason: FocusFailureReason;
-};
-
-/**
- * Describes the message holomush.plugin.v1.FocusFailure.
- * Use `create(FocusFailureSchema)` to create a new message.
- */
-export const FocusFailureSchema: GenMessage<FocusFailure> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 47);
-
-/**
- * PluginHostServiceIsAnyConnFocusedRequest names the character and scene to
- * test for any focused connection.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceIsAnyConnFocusedRequest
- */
-export type PluginHostServiceIsAnyConnFocusedRequest = Message<"holomush.plugin.v1.PluginHostServiceIsAnyConnFocusedRequest"> & {
-  /**
-   * ULID bytes of the character to check.
-   *
-   * @generated from field: bytes character_id = 1;
-   */
-  characterId: Uint8Array;
-
-  /**
-   * ULID bytes of the scene the connections might be focused on.
-   *
-   * @generated from field: bytes scene_id = 2;
-   */
-  sceneId: Uint8Array;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceIsAnyConnFocusedRequest.
- * Use `create(PluginHostServiceIsAnyConnFocusedRequestSchema)` to create a new message.
- */
-export const PluginHostServiceIsAnyConnFocusedRequestSchema: GenMessage<PluginHostServiceIsAnyConnFocusedRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 48);
-
-/**
- * PluginHostServiceIsAnyConnFocusedResponse reports the focus check result.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceIsAnyConnFocusedResponse
- */
-export type PluginHostServiceIsAnyConnFocusedResponse = Message<"holomush.plugin.v1.PluginHostServiceIsAnyConnFocusedResponse"> & {
-  /**
-   * True iff at least one of the character's connections focuses the scene.
-   *
-   * @generated from field: bool focused = 1;
-   */
-  focused: boolean;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceIsAnyConnFocusedResponse.
- * Use `create(PluginHostServiceIsAnyConnFocusedResponseSchema)` to create a new message.
- */
-export const PluginHostServiceIsAnyConnFocusedResponseSchema: GenMessage<PluginHostServiceIsAnyConnFocusedResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 49);
-
-/**
- * PluginHostServiceEvaluateRequest names the action and resource to evaluate.
- * The subject is NOT here — it is recovered host-side from the dispatch token
- * (spec §2, INV-PLUGIN-22).
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceEvaluateRequest
- */
-export type PluginHostServiceEvaluateRequest = Message<"holomush.plugin.v1.PluginHostServiceEvaluateRequest"> & {
-  /**
-   * ABAC action verb to authorize (e.g. "read", "write").
-   *
-   * @generated from field: string action = 1;
-   */
-  action: string;
-
-  /**
-   * resource is a typed instance ref: "scene:01ABC...".
-   *
-   * @generated from field: string resource = 2;
-   */
-  resource: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceEvaluateRequest.
- * Use `create(PluginHostServiceEvaluateRequestSchema)` to create a new message.
- */
-export const PluginHostServiceEvaluateRequestSchema: GenMessage<PluginHostServiceEvaluateRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 50);
-
-/**
- * PluginHostServiceEvaluateResponse returns the ABAC engine's decision.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceEvaluateResponse
- */
-export type PluginHostServiceEvaluateResponse = Message<"holomush.plugin.v1.PluginHostServiceEvaluateResponse"> & {
-  /**
-   * Whether the action is permitted on the resource for the recovered subject.
-   *
-   * @generated from field: bool allowed = 1;
-   */
-  allowed: boolean;
-
-  /**
-   * Human-readable rationale for the decision (e.g. the deny reason).
-   *
-   * @generated from field: string reason = 2;
-   */
-  reason: string;
-
-  /**
-   * Identifier of the policy that produced the decision, when one matched.
-   *
-   * @generated from field: string matched_policy = 3;
-   */
-  matchedPolicy: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceEvaluateResponse.
- * Use `create(PluginHostServiceEvaluateResponseSchema)` to create a new message.
- */
-export const PluginHostServiceEvaluateResponseSchema: GenMessage<PluginHostServiceEvaluateResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 51);
-
-/**
- * PluginHostServiceCommandInfo is per-command metadata returned by ListCommands.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceCommandInfo
- */
-export type PluginHostServiceCommandInfo = Message<"holomush.plugin.v1.PluginHostServiceCommandInfo"> & {
-  /**
-   * name is the canonical command name (e.g. "scene").
-   *
-   * @generated from field: string name = 1;
-   */
-  name: string;
-
-  /**
-   * help is the one-line description from the command registry.
-   *
-   * @generated from field: string help = 2;
-   */
-  help: string;
-
-  /**
-   * usage is the usage pattern (e.g. "scene <subcommand>").
-   *
-   * @generated from field: string usage = 3;
-   */
-  usage: string;
-
-  /**
-   * source is "core" or the owning plugin name.
-   *
-   * @generated from field: string source = 4;
-   */
-  source: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceCommandInfo.
- * Use `create(PluginHostServiceCommandInfoSchema)` to create a new message.
- */
-export const PluginHostServiceCommandInfoSchema: GenMessage<PluginHostServiceCommandInfo> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 52);
-
-/**
- * PluginHostServiceListCommandsRequest names the character whose executable
- * command set to enumerate.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceListCommandsRequest
- */
-export type PluginHostServiceListCommandsRequest = Message<"holomush.plugin.v1.PluginHostServiceListCommandsRequest"> & {
-  /**
-   * character_id is the ULID of the character whose capabilities filter the list.
-   *
-   * @generated from field: string character_id = 1;
-   */
-  characterId: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceListCommandsRequest.
- * Use `create(PluginHostServiceListCommandsRequestSchema)` to create a new message.
- */
-export const PluginHostServiceListCommandsRequestSchema: GenMessage<PluginHostServiceListCommandsRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 53);
-
-/**
- * PluginHostServiceListCommandsResponse returns the filtered command set.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceListCommandsResponse
- */
-export type PluginHostServiceListCommandsResponse = Message<"holomush.plugin.v1.PluginHostServiceListCommandsResponse"> & {
-  /**
-   * commands is the ABAC-filtered set the character may execute.
-   *
-   * @generated from field: repeated holomush.plugin.v1.PluginHostServiceCommandInfo commands = 1;
-   */
-  commands: PluginHostServiceCommandInfo[];
-
-  /**
-   * incomplete is true when engine errors hid some commands from the result.
-   *
-   * @generated from field: bool incomplete = 2;
-   */
-  incomplete: boolean;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceListCommandsResponse.
- * Use `create(PluginHostServiceListCommandsResponseSchema)` to create a new message.
- */
-export const PluginHostServiceListCommandsResponseSchema: GenMessage<PluginHostServiceListCommandsResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 54);
-
-/**
- * PluginHostServiceGetCommandHelpRequest names a command and the character whose
- * access is checked before returning detail.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceGetCommandHelpRequest
- */
-export type PluginHostServiceGetCommandHelpRequest = Message<"holomush.plugin.v1.PluginHostServiceGetCommandHelpRequest"> & {
-  /**
-   * name is the canonical command name to describe.
-   *
-   * @generated from field: string name = 1;
-   */
-  name: string;
-
-  /**
-   * character_id is the ULID of the character whose access is checked.
-   *
-   * @generated from field: string character_id = 2;
-   */
-  characterId: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceGetCommandHelpRequest.
- * Use `create(PluginHostServiceGetCommandHelpRequestSchema)` to create a new message.
- */
-export const PluginHostServiceGetCommandHelpRequestSchema: GenMessage<PluginHostServiceGetCommandHelpRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 55);
-
-/**
- * PluginHostServiceGetCommandHelpResponse returns full help detail.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceGetCommandHelpResponse
- */
-export type PluginHostServiceGetCommandHelpResponse = Message<"holomush.plugin.v1.PluginHostServiceGetCommandHelpResponse"> & {
-  /**
-   * name is the canonical command name.
-   *
-   * @generated from field: string name = 1;
-   */
-  name: string;
-
-  /**
-   * help is the one-line description.
-   *
-   * @generated from field: string help = 2;
-   */
-  help: string;
-
-  /**
-   * usage is the usage pattern.
-   *
-   * @generated from field: string usage = 3;
-   */
-  usage: string;
-
-  /**
-   * help_text is the detailed markdown help body.
-   *
-   * @generated from field: string help_text = 4;
-   */
-  helpText: string;
-
-  /**
-   * source is "core" or the owning plugin name.
-   *
-   * @generated from field: string source = 5;
-   */
-  source: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceGetCommandHelpResponse.
- * Use `create(PluginHostServiceGetCommandHelpResponseSchema)` to create a new message.
- */
-export const PluginHostServiceGetCommandHelpResponseSchema: GenMessage<PluginHostServiceGetCommandHelpResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 56);
-
-/**
- * PluginHostServiceGetSettingRequest reads one owner-partitioned key. The owner
- * is NOT on the wire — the host binds it from the authenticated plugin name
- * (structural cross-plugin isolation).
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceGetSettingRequest
- */
-export type PluginHostServiceGetSettingRequest = Message<"holomush.plugin.v1.PluginHostServiceGetSettingRequest"> & {
-  /**
-   * Scope to read from.
-   *
-   * @generated from field: holomush.plugin.v1.SettingScope scope = 1;
-   */
-  scope: SettingScope;
-
-  /**
-   * Principal ULID: player ID for PLAYER, character ID for CHARACTER, empty for GAME.
-   *
-   * @generated from field: string principal_id = 2;
-   */
-  principalId: string;
-
-  /**
-   * Plugin-owned dot-key to read (e.g. "content.cw_block").
-   *
-   * @generated from field: string key = 3;
-   */
-  key: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceGetSettingRequest.
- * Use `create(PluginHostServiceGetSettingRequestSchema)` to create a new message.
- */
-export const PluginHostServiceGetSettingRequestSchema: GenMessage<PluginHostServiceGetSettingRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 57);
-
-/**
- * PluginHostServiceGetSettingResponse returns a typed list-or-scalar value read
- * from the resolved scope/partition.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceGetSettingResponse
- */
-export type PluginHostServiceGetSettingResponse = Message<"holomush.plugin.v1.PluginHostServiceGetSettingResponse"> & {
-  /**
-   * Whether the key resolved in the requested scope/partition.
-   *
-   * @generated from field: bool found = 1;
-   */
-  found: boolean;
-
-  /**
-   * String-list value (Phase 8 settings are list-valued).
-   *
-   * @generated from field: repeated string string_list = 2;
-   */
-  stringList: string[];
-
-  /**
-   * Scalar string value (for non-list keys).
-   *
-   * @generated from field: string string_value = 3;
-   */
-  stringValue: string;
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceGetSettingResponse.
- * Use `create(PluginHostServiceGetSettingResponseSchema)` to create a new message.
- */
-export const PluginHostServiceGetSettingResponseSchema: GenMessage<PluginHostServiceGetSettingResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 58);
-
-/**
- * PluginHostServiceSetSettingRequest writes one key in the caller's owner
- * partition. Owner is bound host-side, not supplied here.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceSetSettingRequest
- */
-export type PluginHostServiceSetSettingRequest = Message<"holomush.plugin.v1.PluginHostServiceSetSettingRequest"> & {
-  /**
-   * Target scope to write.
-   *
-   * @generated from field: holomush.plugin.v1.SettingScope scope = 1;
-   */
-  scope: SettingScope;
-
-  /**
-   * Principal ULID (empty for GAME).
-   *
-   * @generated from field: string principal_id = 2;
-   */
-  principalId: string;
-
-  /**
-   * Plugin-owned dot-key to write.
-   *
-   * @generated from field: string key = 3;
-   */
-  key: string;
-
-  /**
-   * String-list value to store.
-   *
-   * @generated from field: repeated string string_list = 4;
-   */
-  stringList: string[];
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceSetSettingRequest.
- * Use `create(PluginHostServiceSetSettingRequestSchema)` to create a new message.
- */
-export const PluginHostServiceSetSettingRequestSchema: GenMessage<PluginHostServiceSetSettingRequest> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 59);
-
-/**
- * PluginHostServiceSetSettingResponse is the empty acknowledgement returned on a
- * successful write.
- *
- * @generated from message holomush.plugin.v1.PluginHostServiceSetSettingResponse
- */
-export type PluginHostServiceSetSettingResponse = Message<"holomush.plugin.v1.PluginHostServiceSetSettingResponse"> & {
-};
-
-/**
- * Describes the message holomush.plugin.v1.PluginHostServiceSetSettingResponse.
- * Use `create(PluginHostServiceSetSettingResponseSchema)` to create a new message.
- */
-export const PluginHostServiceSetSettingResponseSchema: GenMessage<PluginHostServiceSetSettingResponse> = /*@__PURE__*/
-  messageDesc(file_holomush_plugin_v1_plugin, 60);
+  messageDesc(file_holomush_plugin_v1_plugin, 13);
 
 /**
  * CommandStatus is the outcome category of a plugin command, mirroring
@@ -2029,107 +692,6 @@ export const AuditEffectSchema: GenEnum<AuditEffect> = /*@__PURE__*/
   enumDesc(file_holomush_plugin_v1_plugin, 1);
 
 /**
- * FocusKind enumerates the types of focused contexts a character can
- * participate in. Adding a new kind requires: (a) a new constant here,
- * (b) a matching session.FocusKind constant in Go, (c) a new
- * FocusKindPolicy implementation registered in the coordinator.
- *
- * @generated from enum holomush.plugin.v1.FocusKind
- */
-export enum FocusKind {
-  /**
-   * Zero value; not a real focus kind — a well-formed FocusKey sets a concrete
-   * kind.
-   *
-   * @generated from enum value: FOCUS_KIND_UNSPECIFIED = 0;
-   */
-  UNSPECIFIED = 0,
-
-  /**
-   * A roleplay scene focus.
-   *
-   * @generated from enum value: FOCUS_KIND_SCENE = 1;
-   */
-  SCENE = 1,
-}
-
-/**
- * Describes the enum holomush.plugin.v1.FocusKind.
- */
-export const FocusKindSchema: GenEnum<FocusKind> = /*@__PURE__*/
-  enumDesc(file_holomush_plugin_v1_plugin, 2);
-
-/**
- * StreamReplayMode controls how a stream subscription's initial replay
- * behaves when added via AddSessionStream.
- *
- * @generated from enum holomush.plugin.v1.StreamReplayMode
- */
-export enum StreamReplayMode {
-  /**
-   * Zero value; treated as FROM_CURSOR for backward compatibility.
-   *
-   * @generated from enum value: STREAM_REPLAY_MODE_UNSPECIFIED = 0;
-   */
-  UNSPECIFIED = 0,
-
-  /**
-   * Replay from the session's saved cursor for the stream (catch-up).
-   *
-   * @generated from enum value: STREAM_REPLAY_MODE_FROM_CURSOR = 1;
-   */
-  FROM_CURSOR = 1,
-
-  /**
-   * Skip historical replay and deliver only newly-arriving events.
-   *
-   * @generated from enum value: STREAM_REPLAY_MODE_LIVE_ONLY = 2;
-   */
-  LIVE_ONLY = 2,
-}
-
-/**
- * Describes the enum holomush.plugin.v1.StreamReplayMode.
- */
-export const StreamReplayModeSchema: GenEnum<StreamReplayMode> = /*@__PURE__*/
-  enumDesc(file_holomush_plugin_v1_plugin, 3);
-
-/**
- * FocusFailureReason is the closed set of per-connection AutoFocusOnJoin
- * failure causes.
- *
- * @generated from enum holomush.plugin.v1.FocusFailureReason
- */
-export enum FocusFailureReason {
-  /**
-   * Zero value; not a real failure reason.
-   *
-   * @generated from enum value: FOCUS_FAILURE_REASON_UNSPECIFIED = 0;
-   */
-  UNSPECIFIED = 0,
-
-  /**
-   * The session lacked the focus membership (JoinFocus was not completed first).
-   *
-   * @generated from enum value: FOCUS_FAILURE_REASON_MEMBERSHIP_ABSENT = 1;
-   */
-  MEMBERSHIP_ABSENT = 1,
-
-  /**
-   * The connection could not be found (e.g. it dropped during the sweep).
-   *
-   * @generated from enum value: FOCUS_FAILURE_REASON_CONNECTION_NOT_FOUND = 2;
-   */
-  CONNECTION_NOT_FOUND = 2,
-}
-
-/**
- * Describes the enum holomush.plugin.v1.FocusFailureReason.
- */
-export const FocusFailureReasonSchema: GenEnum<FocusFailureReason> = /*@__PURE__*/
-  enumDesc(file_holomush_plugin_v1_plugin, 4);
-
-/**
  * SettingScope selects which settings scope a Get/Set targets. There is no
  * chained mode — callers compose scopes themselves (e.g. a CW-block union read
  * across game+player+character). The iokti.7 handler maps each value to its
@@ -2171,14 +733,15 @@ export enum SettingScope {
  * Describes the enum holomush.plugin.v1.SettingScope.
  */
 export const SettingScopeSchema: GenEnum<SettingScope> = /*@__PURE__*/
-  enumDesc(file_holomush_plugin_v1_plugin, 5);
+  enumDesc(file_holomush_plugin_v1_plugin, 2);
 
 /**
  * PluginService is the plugin-IMPLEMENTED half of the binary-plugin contract:
  * the gRPC server runs inside the plugin subprocess (hashicorp/go-plugin) and
  * the host dials it. The host adapter is pkg/plugin.pluginServerAdapter, which
- * bridges these RPCs to a plugin author's Go handler. Contrast PluginHostService
- * below, which the host implements and plugins call back into.
+ * bridges these RPCs to a plugin author's Go handler. The host-implemented
+ * callback surface plugins dial in return is the per-capability
+ * holomush.plugin.host.v1 services (api/proto/holomush/plugin/host/v1).
  *
  * @generated from service holomush.plugin.v1.PluginService
  */
@@ -2243,352 +806,4 @@ export const PluginService: GenService<{
   },
 }> = /*@__PURE__*/
   serviceDesc(file_holomush_plugin_v1_plugin, 0);
-
-/**
- * PluginHostService is the host-IMPLEMENTED half of the contract: the server
- * runs in the host process and binary plugins dial it to call back for event
- * emission, ABAC evaluation, audit decryption, focus mutation, and history
- * reads. Registered live at internal/plugin/goplugin/host_service.go
- * (RegisterPluginHostServiceServer, struct pluginHostServiceServer) — this
- * surface IS served in production.
- *
- * NOTE: the registered server embeds UnimplementedPluginHostServiceServer and
- * implements only 12 of the 18 RPCs below. Log, KVGet, KVSet, KVDelete,
- * AddSessionStream, and RemoveSessionStream are declared here but have no
- * server impl and no production client — they return codes.Unimplemented
- * (tracked in holomush-l6std). Their comments below document that unwired
- * reality, not aspirational behavior.
- *
- * @generated from service holomush.plugin.v1.PluginHostService
- */
-export const PluginHostService: GenService<{
-  /**
-   * EmitEvent publishes one plugin-originated event onto the bus through the
-   * host emit fence. SERVED: pluginHostServiceServer.EmitEvent. The caller's
-   * identity is NOT trusted from the wire — the plugin presents a host-issued
-   * dispatch token in the x-holomush-emit-token metadata header, the host
-   * recovers the vouched-for actor from tokenStore.Lookup(pluginName, token),
-   * and a missing/foreign token is rejected (EMIT_TOKEN_MISSING /
-   * EMIT_TOKEN_REJECTED). The recovered actor then flows through
-   * PluginEventEmitter.Emit, which enforces the manifest gates: emits (subject
-   * namespace must be declared), actor_kinds_claimable (actor kind must be
-   * listed — EMIT_ACTOR_KIND_NOT_CLAIMABLE), and the crypto.emits sensitivity
-   * fence. These gates fire identically for Lua and binary plugins (plugin
-   * runtime symmetry); the token mechanism is the binary-side forgery fence.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.EmitEvent
-   */
-  emitEvent: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceEmitEventRequestSchema;
-    output: typeof PluginHostServiceEmitEventResponseSchema;
-  },
-  /**
-   * Log forwards a plugin log line to the host logger. DECLARED BUT UNSERVED:
-   * pluginHostServiceServer does not implement it and no production client
-   * calls it, so it currently returns codes.Unimplemented (holomush-l6std).
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.Log
-   */
-  log: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceLogRequestSchema;
-    output: typeof PluginHostServiceLogResponseSchema;
-  },
-  /**
-   * KVGet reads a value from the plugin's namespaced key-value store.
-   * DECLARED BUT UNSERVED (holomush-l6std): no server impl, no client; returns
-   * codes.Unimplemented today.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.KVGet
-   */
-  kVGet: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceKVGetRequestSchema;
-    output: typeof PluginHostServiceKVGetResponseSchema;
-  },
-  /**
-   * KVSet writes a value into the plugin's namespaced key-value store.
-   * DECLARED BUT UNSERVED (holomush-l6std): no server impl, no client; returns
-   * codes.Unimplemented today.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.KVSet
-   */
-  kVSet: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceKVSetRequestSchema;
-    output: typeof PluginHostServiceKVSetResponseSchema;
-  },
-  /**
-   * KVDelete removes a key from the plugin's namespaced key-value store.
-   * DECLARED BUT UNSERVED (holomush-l6std): no server impl, no client; returns
-   * codes.Unimplemented today.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.KVDelete
-   */
-  kVDelete: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceKVDeleteRequestSchema;
-    output: typeof PluginHostServiceKVDeleteResponseSchema;
-  },
-  /**
-   * AddSessionStream subscribes an active session to an additional stream
-   * mid-session. DECLARED BUT UNSERVED (holomush-l6std): no server impl, no
-   * production client; returns codes.Unimplemented today. (The wire comment
-   * about SESSION_NOT_FOUND describes the intended-but-unimplemented contract.)
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.AddSessionStream
-   */
-  addSessionStream: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceAddSessionStreamRequestSchema;
-    output: typeof PluginHostServiceAddSessionStreamResponseSchema;
-  },
-  /**
-   * RemoveSessionStream unsubscribes an active session from a stream.
-   * DECLARED BUT UNSERVED (holomush-l6std): no server impl, no production
-   * client; returns codes.Unimplemented today.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.RemoveSessionStream
-   */
-  removeSessionStream: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceRemoveSessionStreamRequestSchema;
-    output: typeof PluginHostServiceRemoveSessionStreamResponseSchema;
-  },
-  /**
-   * JoinFocus adds a focus membership (e.g. a scene) to a session via the
-   * host focus coordinator. SERVED: pluginHostServiceServer.JoinFocus. The
-   * plugin declares intent; the coordinator applies the kind-specific replay
-   * policy. Fails if the focus coordinator is not configured.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.JoinFocus
-   */
-  joinFocus: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceJoinFocusRequestSchema;
-    output: typeof PluginHostServiceJoinFocusResponseSchema;
-  },
-  /**
-   * LeaveFocus removes one focus membership from a session. SERVED:
-   * pluginHostServiceServer.LeaveFocus. Idempotent — leaving a target the
-   * session does not hold is a successful no-op.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.LeaveFocus
-   */
-  leaveFocus: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceLeaveFocusRequestSchema;
-    output: typeof PluginHostServiceLeaveFocusResponseSchema;
-  },
-  /**
-   * LeaveFocusByTarget removes the given focus membership from every
-   * non-expired session that holds it — cross-session fan-out (e.g. a
-   * scene-end reaching all participants). SERVED:
-   * pluginHostServiceServer.LeaveFocusByTarget. Partial success is normal:
-   * per-session failures are aggregated into the response (succeeded /
-   * total_scanned / failed_session_ids) rather than aborting the sweep; the
-   * RPC error is reserved for the enumeration step itself failing.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.LeaveFocusByTarget
-   */
-  leaveFocusByTarget: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceLeaveFocusByTargetRequestSchema;
-    output: typeof PluginHostServiceLeaveFocusByTargetResponseSchema;
-  },
-  /**
-   * PresentFocus repoints a session's PresentingFocus to an existing
-   * membership. SERVED: pluginHostServiceServer.PresentFocus. The target MUST
-   * already be in the session's FocusMemberships (membership is validated, not
-   * implicitly created).
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.PresentFocus
-   */
-  presentFocus: {
-    methodKind: "unary";
-    input: typeof PluginHostServicePresentFocusRequestSchema;
-    output: typeof PluginHostServicePresentFocusResponseSchema;
-  },
-  /**
-   * QueryStreamHistory reads the tail of a stream for plugin-side display,
-   * backward-paginated by opaque cursor. SERVED:
-   * pluginHostServiceServer.QueryStreamHistory via HistoryReader.ReplayTail.
-   * Read-only: it does not advance session cursors or mutate session state.
-   * A negative count is rejected (INVALID_ARGUMENT); count is CLAMPED to
-   * maxQueryStreamHistoryCount (500), not rejected, when too large.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.QueryStreamHistory
-   */
-  queryStreamHistory: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceQueryStreamHistoryRequestSchema;
-    output: typeof PluginHostServiceQueryStreamHistoryResponseSchema;
-  },
-  /**
-   * DecryptOwnAuditRows decrypts a batch of the calling plugin's OWN encrypted
-   * audit rows host-side; the plugin never holds a DEK. SERVED:
-   * pluginHostServiceServer.DecryptOwnAuditRows via ReadbackDecryptor.
-   * DecryptOwnRows. Authorization is two-gate (INV-CRYPTO-27): OwnerMap subject
-   * ownership (g1) plus the crypto.emits[].readback manifest flag (g2). Each
-   * input row gets an independent RowResult (INV-CRYPTO-37) carrying either
-   * plaintext or a stable snake_case no_plaintext_reason ("not_owner",
-   * "auth_guard_deny", "dek_missing", "downgrade_refused", "stale_dek",
-   * "audit_queue_full", "internal"; readback.go reasonToWire). Request /
-   * response shapes and RowResult live in audit.proto (AuditRow domain).
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.DecryptOwnAuditRows
-   */
-  decryptOwnAuditRows: {
-    methodKind: "unary";
-    input: typeof DecryptOwnAuditRowsRequestSchema;
-    output: typeof DecryptOwnAuditRowsResponseSchema;
-  },
-  /**
-   * RequestEmitToken issues a self-token bound to {ActorPlugin, pluginName} so
-   * a plugin-served gRPC handler (e.g. SceneService.CreateScene) — which is NOT
-   * reached via DeliverEvent/DeliverCommand and so holds no dispatch token —
-   * can still call EmitEvent. SERVED: pluginHostServiceServer.RequestEmitToken.
-   * The plugin's identity is taken from the mTLS-bound server struct
-   * (s.pluginName); the request carries no identity fields, so a plugin cannot
-   * impersonate another actor or escalate to a character actor through this
-   * RPC. The actor_kinds_claimable manifest gate still fires when the issued
-   * token is later spent at EmitEvent. (Spec §3.3.5 / §5.4 two-token pattern.)
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.RequestEmitToken
-   */
-  requestEmitToken: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceRequestEmitTokenRequestSchema;
-    output: typeof PluginHostServiceRequestEmitTokenResponseSchema;
-  },
-  /**
-   * SetConnectionFocus is the Phase-5 explicit focus mutation for a single
-   * Connection. SERVED: pluginHostServiceServer.SetConnectionFocus. The
-   * substrate validates the requested membership against the session's
-   * FocusMemberships (D4), then writes Connection.FocusKey and (D9-gated)
-   * Info.PresentingFocus atomically under one Store-lock acquisition (D7).
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.SetConnectionFocus
-   */
-  setConnectionFocus: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceSetConnectionFocusRequestSchema;
-    output: typeof PluginHostServiceSetConnectionFocusResponseSchema;
-  },
-  /**
-   * GetConnectionFocus returns the named connection's current per-connection
-   * focus, or absent when the connection is grid-focused (FocusKey nil) or
-   * unknown. Read-only counterpart of SetConnectionFocus; lets plugins route
-   * connection-scoped operations (e.g. scene pose) to the focused target.
-   * See goplugin/host_service.go::GetConnectionFocus.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.GetConnectionFocus
-   */
-  getConnectionFocus: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceGetConnectionFocusRequestSchema;
-    output: typeof PluginHostServiceGetConnectionFocusResponseSchema;
-  },
-  /**
-   * AutoFocusOnJoin is the Phase-5 fan-out that focuses all of a character's
-   * terminal/telnet connections on a scene at once. SERVED:
-   * pluginHostServiceServer.AutoFocusOnJoin. Connections already explicitly
-   * focused elsewhere are skipped (D8). The caller MUST have completed
-   * JoinFocus first, since the substrate requires the membership to exist.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.AutoFocusOnJoin
-   */
-  autoFocusOnJoin: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceAutoFocusOnJoinRequestSchema;
-    output: typeof PluginHostServiceAutoFocusOnJoinResponseSchema;
-  },
-  /**
-   * IsAnyConnFocused is the Phase-5 notification-emission helper: it reports
-   * whether any of the character's connections currently focuses the given
-   * scene, so callers can decide whether to emit a focus-related notification.
-   * SERVED: pluginHostServiceServer.IsAnyConnFocused.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.IsAnyConnFocused
-   */
-  isAnyConnFocused: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceIsAnyConnFocusedRequestSchema;
-    output: typeof PluginHostServiceIsAnyConnFocusedResponseSchema;
-  },
-  /**
-   * Evaluate runs the host ABAC engine for one action against one resource
-   * instance owned by the calling plugin. SERVED:
-   * pluginHostServiceServer.Evaluate. The subject is derived host-side from the
-   * dispatch token exactly as EmitEvent does (token→actor recovery) — there is
-   * no subject field on the wire (spec §2, INV-PLUGIN-22). Fails closed on nil engine,
-   * missing/rejected token, empty actor subject, or a resource type the plugin
-   * does not own.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.Evaluate
-   */
-  evaluate: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceEvaluateRequestSchema;
-    output: typeof PluginHostServiceEvaluateResponseSchema;
-  },
-  /**
-   * ListCommands enumerates the commands the named character may execute,
-   * ABAC-filtered by the host. SERVED: pluginHostServiceServer.ListCommands,
-   * delegating to commandquery.Querier.Available. The subject is the request's
-   * character_id (parity with the Lua holomush.list_commands(character_id) host
-   * function — not the dispatch-token actor, since this is read-only metadata,
-   * not an actor-gated mutation). incomplete is true when engine errors hid
-   * some commands.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.ListCommands
-   */
-  listCommands: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceListCommandsRequestSchema;
-    output: typeof PluginHostServiceListCommandsResponseSchema;
-  },
-  /**
-   * GetCommandHelp returns full help detail for one command after an access
-   * check for character_id. SERVED: pluginHostServiceServer.GetCommandHelp,
-   * delegating to commandquery.Querier.Help. Mirrors the Lua
-   * holomush.get_command_help(name, character_id) host function.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.GetCommandHelp
-   */
-  getCommandHelp: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceGetCommandHelpRequestSchema;
-    output: typeof PluginHostServiceGetCommandHelpResponseSchema;
-  },
-  /**
-   * GetSetting reads a single-scope setting in the calling plugin's owner
-   * partition (owner bound host-side from the authenticated plugin name, never
-   * from the request). The handler resolves SettingScope to its backing store;
-   * a missing key returns a successful response with found=false, never a
-   * codes.NotFound status error.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.GetSetting
-   */
-  getSetting: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceGetSettingRequestSchema;
-    output: typeof PluginHostServiceGetSettingResponseSchema;
-  },
-  /**
-   * SetSetting writes a single-scope setting in the calling plugin's partition;
-   * GAME scope requires an operator authorization decision (host-enforced, not
-   * trusted from the wire). The owner partition is bound host-side from the
-   * authenticated plugin name.
-   *
-   * @generated from rpc holomush.plugin.v1.PluginHostService.SetSetting
-   */
-  setSetting: {
-    methodKind: "unary";
-    input: typeof PluginHostServiceSetSettingRequestSchema;
-    output: typeof PluginHostServiceSetSettingResponseSchema;
-  },
-}> = /*@__PURE__*/
-  serviceDesc(file_holomush_plugin_v1_plugin, 1);
